% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_remote_data.R
\name{access_remote_file}
\alias{access_remote_file}
\title{Access Remote Data from PerfettoLab Server}
\usage{
access_remote_file(file, dir)
}
\arguments{
\item{file}{Character. The name of the file to download (e.g. "example.tsv").}

\item{dir}{Character. The subdirectory within the SignalingProfiler directory where the file is located (e.g. "PKN").}
}
\value{
A tibble containing the parsed contents of the downloaded TSV file.
}
\description{
Downloads and reads a `.tsv` file hosted on the PerfettoLab server 
(https://perfettolab.bio.uniroma1.it/PerfettoLabData/SignalingProfiler).
This function is useful for accessing publicly available signaling datasets 
for downstream analyses in the context of SignalingProfiler.
}
\examples{
file <- "SerThrAtlas_PsP_regulons_filtered_99.tsv"
access_remote_file(file = file, dir = "PKN")

}
