% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{add_output_carnival_edges_attributes}
\alias{add_output_carnival_edges_attributes}
\title{add_output_carnival_edges_attributes}
\usage{
add_output_carnival_edges_attributes(carnival_result)
}
\arguments{
\item{carnival_result}{runCARNIVAL output}
}
\value{
dataframe with edges attributes
}
\description{
add_output_carnival_edges_attributes
}
