% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{add_output_carnival_nodes_attributes}
\alias{add_output_carnival_nodes_attributes}
\title{Title}
\usage{
add_output_carnival_nodes_attributes(
  carnival_result,
  proteins_df,
  organism,
  prot_df
)
}
\arguments{
\item{carnival_result}{runCARNIVAL output}

\item{proteins_df}{inferred proteins df}

\item{organism}{string, 'mouse' or 'human'}

\item{prot_df}{proteomics data}
}
\value{
dataframe with nodes and attributes
}
\description{
Title
}
