% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background_phenoscore}
\alias{background_phenoscore}
\title{SIGNOR proteins as background for PhenoScore Randomization}
\format{
An object of class \code{character} of length 9987.
}
\usage{
background_phenoscore
}
\description{
Vector of Gene Symbols from SIGNOR database, generated with [sp_parsing]
function; from this dataset a random set of proteins is sampled in
PhenoScore algorithm analysis.
}
\keyword{datasets}
