% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{check_CARNIVAL_inputs}
\alias{check_CARNIVAL_inputs}
\title{check_CARNIVAL_inputs}
\usage{
check_CARNIVAL_inputs(
  source_df,
  target_df,
  naive_network,
  proteins_df,
  organism
)
}
\arguments{
\item{source_df}{tibble with source nodes discretized among 1 and -1}

\item{target_df}{tibble with target nodes in a continuous range of activity}

\item{naive_network}{tibble with naive network in SIF format}

\item{proteins_df}{all inferred proteins df}

\item{organism}{string, organism you are using}
}
\value{
TRUE if everything is correct
}
\description{
check_CARNIVAL_inputs
}
