% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_PKN.R
\name{choose_PKN}
\alias{choose_PKN}
\title{choose_PKN}
\usage{
choose_PKN(organism, with_atlas = FALSE, custom = FALSE, custom_path = NULL)
}
\arguments{
\item{organism}{string, 'mouse' or 'human'}

\item{with_atlas}{Boolean value, default FALSE, if TRUE you can use PKN integrated with kinome atlas data}

\item{custom}{Boolean value, default FALSE, if TRUE you can provide the path for your custom PKN}

\item{custom_path}{string, specify the path of the custom PKN in tab separated format}
}
\value{
PKN_table, table of PKN
}
\description{
choose_PKN
}
