% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore.R
\name{compute_phenoscore}
\alias{compute_phenoscore}
\title{compute_phenoscore}
\usage{
compute_phenoscore(
  sp_output,
  desired_phenotypes,
  compact = TRUE,
  max_path_length = 4,
  remove_cascade = TRUE,
  use_carnival_activity = FALSE
)
}
\arguments{
\item{sp_output}{Signaling Profiler output list}

\item{desired_phenotypes}{NULL if you want all phenotypes, vector of phenotypes name in uppercase}

\item{compact}{TRUE if you want a compact result, FALSE otherwise}

\item{max_path_length}{integer, from 1 to 4, path length of}

\item{remove_cascade}{default TRUE}

\item{use_carnival_activity}{Boolean value, TRUE uses all proteins in network,
FALSE only experimentally derived proteins}
}
\value{
phenoSCORE table
}
\description{
compute_phenoscore
}
