% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{concatenate_molecular_functions}
\alias{concatenate_molecular_functions}
\title{concatenate_molecular_functions}
\usage{
concatenate_molecular_functions(prediction_output_list)
}
\arguments{
\item{prediction_output_list}{list with three df for molecular functions}
}
\value{
concantantion of dfs
}
\description{
concatenate_molecular_functions
}
