% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore.R
\name{config_env}
\alias{config_env}
\title{config_env}
\usage{
config_env(local = FALSE)
}
\arguments{
\item{local}{Boolean, default is FALSE (for development purposes)}
}
\description{
This function configurates SignalingProfiler_env for securely running python
}
