% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{coverage_of_inferred_proteins_in_db}
\alias{coverage_of_inferred_proteins_in_db}
\title{coverage_of_inferred_proteins_in_db}
\usage{
coverage_of_inferred_proteins_in_db(prediction_output, organism, report)
}
\arguments{
\item{prediction_output}{concatenation of inferred proteins from omics data}

\item{organism}{'human' or 'mouse'}

\item{report}{' report of coverage analysis}
}
\value{
print information about coverage or annotates them in a file
}
\description{
coverage_of_inferred_proteins_in_db
}
