% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{create_discretized_initiators_for_carnival}
\alias{create_discretized_initiators_for_carnival}
\title{create_discretized_initiators_for_carnival}
\usage{
create_discretized_initiators_for_carnival(proteins_df)
}
\arguments{
\item{proteins_df}{initiators protein df}
}
\value{
initiators df with activities discretized
}
\description{
create_discretized_initiators_for_carnival
}
