% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoscore.R
\name{create_fasta}
\alias{create_fasta}
\title{Create a FASTA File from Phosphoproteomics Data}
\usage{
create_fasta(phospho_df, path)
}
\arguments{
\item{phospho_df}{A dataframe containing phosphoproteomics data.}

\item{path}{A string specifying the output path of the FASTA file.}
}
\value{
This function writes a FASTA file to the specified path and returns `NULL`.
}
\description{
This function generates a FASTA file from phosphoproteomics data, which can be used for sequence alignment.
The file has the following format:
>MAPK1-XXXXXXXXSXXXXXXX-S-15-mouse
XXXXXXXXSXXXXXXX
}
\examples{
# Example phosphoproteomics dataset
phospho_data <- data.frame(
  gene_name = c("TP53", "MYC"),
  UNIPROT = c("P04637", "P01106"),
  sequence_window = c("XXXXXXXXSXXXXXXX", "XXXXXXXTXXXXXXXX"),
  aminoacid = c("S", "T"),
  position = c(15, 22)
)

# Create a FASTA file
create_fasta(phospho_data, "output.fasta")

}
