% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_information.R
\name{create_ksea_regulons}
\alias{create_ksea_regulons}
\title{create_ksea_regulons}
\usage{
create_ksea_regulons(
  resources = c("SIGNOR", "Omnipath"),
  organism,
  omni_resources = c("PhosphoSite", "PhosphoSite_ProtMapper", "SIGNOR",
    "SIGNOR_ProtMapper", "Reactome_ProtMapper", "phosphoELM", "phosphoELM", "HPRD",
    "DEPOD")
)
}
\arguments{
\item{resources}{vector, sources such as SIGNOR or OmniPath}

\item{organism}{string, 'human' or 'mouse'}

\item{omni_resources}{vector, see OmniPath tutorial for a complete list of kin-substrates resources}
}
\value{
a table of KIN/PHOS-target sites from user-defined OmniPath sources
}
\description{
create_ksea_regulons
}
