% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{create_naive_network_construction_input}
\alias{create_naive_network_construction_input}
\title{create_naive_network_construction_input}
\usage{
create_naive_network_construction_input(prediction_output, mf)
}
\arguments{
\item{prediction_output}{output of activity prediction pipeline}

\item{mf}{molecular function of the proteins in proteins' table}
}
\value{
output of activity prediction pipeline with just activity
}
\description{
create_naive_network_construction_input
}
