% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_prediction.R
\name{create_viper_format}
\alias{create_viper_format}
\title{Create VIPER-compatible dataset}
\usage{
create_viper_format(omic_data, analysis, significance)
}
\arguments{
\item{omic_data}{A dataframe containing phosphoproteomics or transcriptomics data.}

\item{analysis}{A string indicating the type of analysis to perform. Options are:
- "tfea": Transcription Factor Enrichment Analysis (TFEA).
- "ksea": Kinase-Substrate Enrichment Analysis (KSEA).}

\item{significance}{A logical value indicating whether to filter for significant analytes.
If `TRUE`, only significant analytes (where `significant == '+'`) are included.}
}
\value{
A data frame formatted for VIPER, containing columns:
  - `ID`: Gene name (TFEA) or phosphosite ID (KSEA).
  - `logFC`: Fold change value.
  - `t`: T-statistic for analysis.
  - `adj.P`: Adjusted p-value.
}
\description{
This function converts phosphoproteomics or transcriptomics data into a format
suitable for VIPER (Virtual Inference of Protein-activity by Enriched Regulon analysis)
(PMID:  27322546).
}
\examples{
# Example dataset with transcriptomics data
omic_data <- data.frame(
  gene_name = c("TP53", "MYC", "EGFR"),
  difference = c(1.5, -2.1, 0.8),
  logpval = c(0.01, 0.05, 0.2),
  significant = c("+", "+", NA)
)

# Format for TFEA with only significant genes
formatted_data_sign <- create_viper_format(omic_data, analysis = "tfea", significance = TRUE)

# Format for TFEA including all analytes
formatted_data_all <- create_viper_format(omic_data, analysis = "tfea", significance = FALSE)

}
