% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_PKN}
\alias{db_PKN}
\alias{db_human_dir}
\alias{db_human_ind}
\alias{db_human_atlas_dir}
\alias{db_human_atlas_ind}
\alias{db_mouse_dir}
\alias{db_mouse_ind}
\title{Prior Knowledge Networks (PKN) as igraph Objects}
\format{
Each `db_*` object is an `igraph` object with:
- **Vertex attributes**:
  \describe{
    \item{name}{Character, Biological Entity Name, corresponds to `ENTITY` column of tabular form}
    \item{ID}{Character, Biological Entity Identifier}
    \item{TYPE}{Character, Biological Entity Molecular Type}
    \item{DATABASE}{Character, Database Source for Biological Entity Identifier}
  }
- **Edge attributes**:
  \describe{
    \item{INTERACTION}{Character, Sign of the Interaction, 1 Activating, -1 Inhibiting}
    \item{IDA}{Character, Source Biological Entity Identifier}
    \item{IDB}{Character, Target Biological Entity Identifier}
    \item{MECHANISM}{Character, Molecular Mechanism of the Interaction}
    \item{RESIDUE}{Character, if the MECHANISM is a phosphorylation, the modified residue in the interaction}
    \item{SEQUENCE}{Character, if the MECHANISM is a phosphorylation, a 15mer centered on the modified residue}
    \item{TYPEA}{Character, Source Biological Entity Molecular Type}
    \item{DATABASEA}{Character, Source Biological Entity Database}
    \item{TYPEB}{Character, Target Biological Entity Molecular Type}
    \item{DATABASEB}{Character, Target Biological Entity Database}
    \item{PHOSPHO_KEY_GN_SEQ}{Character composed of ENTITYB and SEQUENCE column}
    \item{SOURCE}{Character, Source Database for the interaction; either SIGNOR or PhosphoSitePlus or both}
    \item{DIRECT}{Logical, whether the interaction is direct}
  }

An object of class \code{igraph} of length 9692.

An object of class \code{igraph} of length 9675.

An object of class \code{igraph} of length 9744.

An object of class \code{igraph} of length 8405.

An object of class \code{igraph} of length 8455.
}
\source{
- <https://signor.uniroma2.it>
- <https://www.phosphosite.org>
- <https://pubmed.ncbi.nlm.nih.gov/36631611/> (for kinase-substrate interactions)
- <https://pubmed.ncbi.nlm.nih.gov/38720073/> (for kinase-substrate interactions)
}
\usage{
db_human_dir

db_human_ind

db_human_atlas_dir

db_human_atlas_ind

db_mouse_dir

db_mouse_ind
}
\description{
Directed interactomes represented as `igraph` objects, reporting processed interactions
from SIGNOR and PhosphoSitePlus. Some versions include *experimentally-derived*
kinase-substrate relationships from Serine, Threonine, and Tyrosine Atlases (PMIDs: 36631611, 38720073).
}
\details{
Available datasets:
- `db_human_dir`: Human Prior Knowledge Network with *direct* interactions.
- `db_human_ind`: Human Prior Knowledge Network with *direct* and *indirect* interactions.
- `db_human_atlas_dir`: Human PKN with *direct* interactions and kinase-substrate relationships.
- `db_human_atlas_ind`: Human PKN with *direct* and *indirect* interactions and kinase-substrate relationships.
- `db_mouse_dir`: Mouse Prior Knowledge Network with *direct* interactions.
- `db_mouse_ind`: Mouse Prior Knowledge Network with *direct* and *indirect* interactions.
}
\keyword{datasets}
