% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{default_CARNIVAL_options}
\alias{default_CARNIVAL_options}
\title{default_CARNIVAL_options}
\usage{
default_CARNIVAL_options(solver = NULL)
}
\arguments{
\item{solver}{string specifying solver}
}
\value{
list of CARNIVAL options
}
\description{
default_CARNIVAL_options
}
