% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{expand_and_map_edges}
\alias{expand_and_map_edges}
\title{expand_and_map_edges}
\usage{
expand_and_map_edges(
  optimized_graph_rds,
  optimized_graph_sif,
  organism,
  phospho_df,
  files,
  path_sif,
  path_rds
)
}
\arguments{
\item{path_rds}{}
}
\description{
expand_and_map_edges
}
