% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_carnival_edges}
\alias{extract_carnival_edges}
\title{Extract Edges from an iGraph Object}
\usage{
extract_carnival_edges(graph)
}
\arguments{
\item{graph}{An `igraph` object representing a network.}
}
\value{
A tibble where each row represents an edge, containing at least
  the `source` and `target` nodes.
}
\description{
This function extracts edge information from an `igraph` network
and returns it as a tibble.
}
\examples{
library(igraph)
data('toy_opt_network')
extract_carnival_edges(toy_opt_network$igraph_network)
}
