% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{filter_inferred_protein_for_presence_in_naive_network}
\alias{filter_inferred_protein_for_presence_in_naive_network}
\title{filter_inferred_protein_for_presence_in_naive_network}
\usage{
filter_inferred_protein_for_presence_in_naive_network(
  naive_network,
  prediction_output
)
}
\arguments{
\item{naive_network}{naive network connecting inferred proteins}

\item{prediction_output}{inferred proteins from experimental data}
}
\value{
inferred protein filtered for presence in naive network
}
\description{
filter_inferred_protein_for_presence_in_naive_network
}
