% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{format_for_visualization}
\alias{format_for_visualization}
\title{format_for_visualization}
\usage{
format_for_visualization(sp_object)
}
\arguments{
\item{sp_object}{list of nodes, edges and graph of SP}
}
\value{
sp_object with attributes on nodes and edges for Cytoscape visualization
}
\description{
format_for_visualization
}
