% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{formatting_proteins_for_carnival}
\alias{formatting_proteins_for_carnival}
\title{formatting_proteins_for_carnival}
\usage{
formatting_proteins_for_carnival(proteins_df)
}
\arguments{
\item{proteins_df}{dataframe with inferred proteins with specific mf}
}
\value{
a list of proteins as desired by CARNIVAL
}
\description{
formatting_proteins_for_carnival
}
