% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{generate_hybrid_db}
\alias{generate_hybrid_db}
\title{Generates hybrid regulatory db}
\usage{
generate_hybrid_db(mh_alignment, activatory)
}
\arguments{
\item{mh_alignment}{blastp output alignment among mouse and human}

\item{activatory}{boolean value}
}
\value{
hybrid database
}
\description{
Generates hybrid regulatory db
}
