% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{generateTFList}
\alias{generateTFList}
\title{generateTFList}
\usage{
generateTFList(df = df, top = 50, access_idx = 1)
}
\arguments{
\item{df}{df of inferred proteins}

\item{top}{integer}

\item{access_idx}{integer}
}
\value{
list of proteins
}
\description{
generateTFList
}
