% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{get_all_edge_ids}
\alias{get_all_edge_ids}
\title{get all edge ids}
\usage{
get_all_edge_ids(all_shortest_paths, db)
}
\arguments{
\item{all_shortest_paths}{takes in input all shortest paths of igraph command}

\item{db}{fb used for the analysis}
}
\value{
all edges id along the shortest path
}
\description{
get all edge ids
}
