% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{get_all_one_step_connections}
\alias{get_all_one_step_connections}
\title{get_all_one_step_connections}
\usage{
get_all_one_step_connections(
  source_gn,
  target_gn,
  proteins_df,
  db,
  rds_path = "direct_naive.RDS",
  sif_path = "direct_naive.sif"
)
}
\arguments{
\item{proteins_df}{all inferred proteins}

\item{db}{built in causal relation database}

\item{source_df}{gene names of source nodes}

\item{target_df}{gene names of target nodes}
}
\value{
naive network
}
\description{
get_all_one_step_connections
}
