% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_paths_custom.R
\name{get_all_shortest_path_custom}
\alias{get_all_shortest_path_custom}
\title{get_all_shortest_path_custom}
\usage{
get_all_shortest_path_custom(
  start_nodes_gn,
  target_nodes_gn,
  organism,
  max_length
)
}
\arguments{
\item{start_nodes_gn}{vector of gene_names of starting nodes}

\item{target_nodes_gn}{vector of gene_names of target nodes}

\item{organism}{string, 'mouse' or 'human'}

\item{max_length}{integer, 1 to 4}
}
\description{
get_all_shortest_path_custom
}
