% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_information.R
\name{get_omnipath}
\alias{get_omnipath}
\title{get_omnipath}
\usage{
get_omnipath(
  organism,
  resources = c("PhosphoSite", "PhosphoSite_ProtMapper", "SIGNOR", "SIGNOR_ProtMapper",
    "Reactome_ProtMapper", "phosphoELM", "phosphoELM", "HPRD", "DEPOD")
)
}
\arguments{
\item{organism}{string, 'human' or 'mouse'}

\item{resources}{vector, see OmniPath tutorial for a complete list of kin-substrates resources}
}
\value{
a table of KIN/PHOS-target sites from user-defined OmniPath sources
}
\description{
get_omnipath
}
