% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_phosphosites_creation.R
\name{get_phosphoscore_info}
\alias{get_phosphoscore_info}
\title{Retrieve PhosphoScore Information}
\usage{
get_phosphoscore_info(
  resources = c("SIGNOR", "PsP"),
  organism,
  psp_reg_site_path = NULL,
  psp_kin_sub_path = NULL,
  only_activatory = TRUE,
  aa_pos = FALSE,
  local = FALSE
)
}
\arguments{
\item{resources}{Character vector specifying the regulatory phosphosite databases to query. Allowed values are `"SIGNOR"` and `"PsP"`.}

\item{organism}{Character string specifying the organism. Supported values are `"human"` and `"mouse"`.}

\item{psp_reg_site_path}{Character string specifying the file path to the PhosphoSitePlus 'Regulatory Sites' data.
Required if `"PsP"` is included in `resources`.}

\item{psp_kin_sub_path}{Character. Path to the PhosphoSitePlus "Kinases_Substrate_Dataset" file manually downloaded from it.}

\item{only_activatory}{Logical, indicating whether to filter for only phosphosites that regulate protein activity. Default is `TRUE`.}

\item{aa_pos}{Logical, whether using in combination with Gene Symbol the amino acid residue position (`RESIDUE`)
instead of 15-mer peptide (`SEQUENCE`) for phosphosite identification. Default is `FALSE`.}

\item{local}{Logical, indicating whether to use a local or package-installed version of the SIGNOR dictionary. Default is `FALSE`.}
}
\value{
A tibble containing regulatory phosphosites with the following columns:
- `PHOSPHO_KEY_GN_SEQ`: Unique phosphosite identifier (`Gene-Residue` or `Gene-Sequence`).
- `UNIPROT`: UniProt accession ID.
- `ACTIVATION`: Regulatory effect (`1` for activation, `-1` for inhibition).
}
\description{
This function compiles phosphorylation site regulatory data from SIGNOR and PhosphoSitePlus (PsP).
The resulting dataset identifies sites with consistent activation or inhibition effects.
}
\details{
- If `"SIGNOR"` is included in `resources`, the function extracts phosphorylation site interactions from SIGNOR.
- If `"PsP"` is included, the function extracts phosphorylation sites from PhosphoSitePlus.
- The function removes phosphosites with conflicting regulatory effects (e.g., activating in one context but inhibiting in another).
- If `aa_pos = TRUE`, phosphosites are identified using their residue position (e.g., `"MAPK1-T185"`),
  otherwise, their phosphopeptide context (e.g., `ABL1-RLMTGDTYTAHAGAK `) is used.

- If `"SIGNOR"` is included in `resources`, the function extracts phosphorylation site interactions from SIGNOR.
- If `"PsP"` is included, the function extracts phosphorylation sites from PhosphoSitePlus.
- The function removes phosphosites with conflicting regulatory effects (e.g., activating in one context but inhibiting in another).
}
\examples{

# Retrieve phosphosite information from SIGNOR
phosphoscore_info <- get_phosphoscore_info(resources = c("SIGNOR"),
                                           organism = "human")

\dontrun{
# Retrieve phosphosite information from SIGNOR and PsP for human using sequence-based identifiers
phosphoscore_info <- get_phosphoscore_info(resources = c("SIGNOR", "PsP"),
                                           organism = "human",
                                           psp_reg_site_path = "./Regulatory_sites")

# Retrieve phosphosite information using amino acid residue position instead of sequence
phosphoscore_info_residue <- get_phosphoscore_info(resources = c("SIGNOR", "PsP"),
                                                   organism = "human",
                                                   psp_reg_site_path = "./Regulatory_sites",
                                                   aa_pos = TRUE)
}

}
