% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_information.R
\name{get_signor_regulons}
\alias{get_signor_regulons}
\title{get_signor_regulons}
\usage{
get_signor_regulons(organism, analysis)
}
\arguments{
\item{organism}{string, 'human' and 'mouse'}

\item{analysis}{string, 'tfea' and 'ksea'}
}
\value{
table of regulons derived from SIGNOR
}
\description{
get_signor_regulons
}
