% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{igraphToSif}
\alias{igraphToSif}
\title{Convert an igraph Object to a SIF (Simple Interaction Format) File}
\usage{
igraphToSif(inGraph, outfile = "output.sif", edgeLabel = "label")
}
\arguments{
\item{inGraph}{An `igraph` object representing the network.}

\item{outfile}{A string specifying the output file path for the SIF file; default: `"output.sif"`}

\item{edgeLabel}{A string specifying the edge attribute to use as an interaction label in the SIF file; default: `"label"`}
}
\value{
`NULL`. The function creates a SIF file named as defined by `outfile`.
}
\description{
This function exports an `igraph` object into a SIF (Simple Interaction Format) file,
commonly used for representing network interactions.
}
\details{
The SIF format represents network interactions using a three-column structure:
\tabular{lll}{
  **Node1** \tab **Interaction Type** \tab **Node2** \cr
}
Each row represents an interaction between two nodes, with the interaction type
taken from the specified edge attribute (`edgeLabel`).
If a file with the same name as `outfile` already exists, it will be overwritten.
}
\examples{

library(igraph)
g <- graph_from_edgelist(matrix(c("A", "B", "B", "C"), byrow=TRUE, ncol=2))
E(g)$label <- c(1, -1)
igraphToSif(g, outfile="network.sif", edgeLabel="label")

}
