% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sp_py.R
\name{install_sp_py}
\alias{install_sp_py}
\title{Install Python Dependencies for SignalingProfiler}
\usage{
install_sp_py(
  envname = "r-signalingprofiler",
  new_env = identical(envname, "r-signalingprofiler")
)
}
\arguments{
\item{envname}{"r-signalingprofiler"}

\item{new_env}{Logical. If `TRUE` (default when `envname` is `"r-signalingprofiler"`),
the existing virtual environment will be removed before installation.}
}
\value{
This function does not return a value; it installs Python dependencies.
}
\description{
This function installs the required Python dependencies for
the SignalingProfiler package within a specified virtual environment.
If the default environment name ("r-signalingprofiler") is used and it already exists,
the function removes and recreates it before installing dependencies.
}
\details{
his function ensures that the required Python dependencies (`requests` and `pandas`)
are installed in the specified virtual environment. If `new_env` is `TRUE`,
the function first removes any existing virtual environment with the same name before proceeding.
}
\examples{
install_sp_py()
install_sp_py(envname = "custom-env", new_env = FALSE)
}
