% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_information.R
\name{integrate_atlas_in_pkn}
\alias{integrate_atlas_in_pkn}
\title{integrate_atlas_in_pkn}
\usage{
integrate_atlas_in_pkn(
  reg_site_path,
  kin_sub_path,
  file_path = NULL,
  local = FALSE
)
}
\arguments{
\item{reg_site_path}{path to the PhosphoSitePlus 'Regulatory Sites' file}

\item{kin_sub_path}{path to the PhosphoSitePlus 'Kinases-Substrates' file}

\item{file_path}{string, path for file saving}
}
\value{
table of interactions with Ser/Thr-derived interactions included
}
\description{
integrate_atlas_in_pkn
}
