% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{map_experimental_on_regulatory_phosphosites}
\alias{map_experimental_on_regulatory_phosphosites}
\title{Title}
\usage{
map_experimental_on_regulatory_phosphosites(
  phosphoproteomic_data,
  organism,
  activatory,
  path_fasta,
  blastp_path = NULL,
  local = FALSE,
  custom = FALSE,
  custom_path = NULL
)
}
\arguments{
\item{phosphoproteomic_data}{dataset of phosphoproteomics data}

\item{organism}{string, human mouse or hybrid}

\item{activatory}{boolean value, if relation activatory or all}

\item{blastp_path}{path of blastp in windows}

\item{local}{DEVELOPMENTAL PURPOSES TRUE OR FALSE}

\item{custom}{Boolean, default FALSE, TRUE if you want to provide your regulatory phosphosites}

\item{custom_path}{string to the .tsv file of custom regulatory phosphosites}
}
\value{
phosphoscore_df representing experimentally quantified phosphosites
associated to their fold-change
}
\description{
Title
}
