% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_term_managment.R
\name{molecular_function_annotation}
\alias{molecular_function_annotation}
\title{Annotate Proteins with GO Molecular Function}
\usage{
molecular_function_annotation(inferred_proteins_dataset, organism)
}
\arguments{
\item{inferred_proteins_dataset}{Data frame, protein table containing a `gene_name` column.}

\item{organism}{Character, organism type. Must be one of `"human"` or `"mouse"`.}
}
\value{
A data frame identical to `inferred_proteins_dataset`, but with an
additional `mf` column containing GO Molecular Function annotations:
  - `"rec"` if the protein is user-defined.
  - `"kin"` if the protein is a kinase.
  - `"phos"` if the protein is a phosphatase.
  - `"tf"` if the protein is a transcription factor.
  - `"other"` for proteins without a matching annotation.
  - `"complex"` for proteins labeled with `SIGNOR-C`.
  - `"fusion protein"` for proteins labeled with `SIGNOR-F`.
}
\description{
Annotates a table of proteins with Gene Ontology (GO) Molecular Function
categories based on a static reference table (`gomf_human` or `gomf_mouse`).
The annotation is performed using the `gene_name` column as the key.
}
\details{
The function joins the input dataset with preloaded molecular function
tables (`gomf_human` or `gomf_mouse`) using the `gene_name` column.
If an `mf` column is already present, it is removed before annotation.
}
\examples{
data('toy_prot_activity_df')
molecular_function_annotation(
  inferred_proteins_dataset = toy_prot_activity_df,
  organism = "human")

}
