% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{one_layer_naive_network}
\alias{one_layer_naive_network}
\title{one_layer_naive_network}
\usage{
one_layer_naive_network(
  receptors_gn,
  targets_gn,
  db,
  rds_path = "one_layer_naive.RDS",
  sif_path = "one_layer_naive.sif"
)
}
\arguments{
\item{receptors_gn}{gene names of receptors}

\item{targets_gn}{gene names of target nodes (transcription factors)}

\item{db}{built in causal relation database}
}
\value{
naive network
}
\description{
one_layer_naive_network
}
