% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_functional_circuit.R
\name{pheno_to_start_circuit}
\alias{pheno_to_start_circuit}
\title{Extract QueryNodes-to-Phenotype Circuit from SignalingProfiler Network}
\usage{
pheno_to_start_circuit(
  SP_object,
  start_nodes,
  phenotypes,
  k,
  start_to_top = FALSE
)
}
\arguments{
\item{SP_object}{A list containing the SignalingProfiler network (igraph object), nodes table, and edges table.}

\item{start_nodes}{A character vector of nodes representing the start points in the model.}

\item{phenotypes}{A character vector of phenotype names to be linked to the start nodes.}

\item{k}{An integer specifying the maximum path length between phenotypes and start nodes.}

\item{start_to_top}{Logical, whether to remove incoming edges to start nodes (`TRUE` or `FALSE`, default: `FALSE`).}
}
\value{
An `igraph` object representing the extracted subnetwork with paths from phenotypes to start nodes.
}
\description{
This function extracts a subnetwork linking specified start nodes
to phenotypes within a given path length (`k`). It generates a
directed subgraph where only paths from phenotypes to start nodes
are retained.
}
\details{
The function:
- Identifies shortest paths from each phenotype to the defined start nodes.
- Retains paths that are at most `k` steps long.
- Extracts a subgraph containing the identified paths.
- (Optional) Removes incoming edges to start nodes if `start_to_top = TRUE`.

If no valid paths are found, a warning is issued.
}
\examples{
data('toy_sp_output')
pheno_to_start_circuit(SP_object = toy_sp_output, start_nodes = 'FLT3', phenotypes = c('PROLIFERATION', 'APOPTOSIS'), k = 4)

}
