% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore.R
\name{phenoscore_computation_v1}
\alias{phenoscore_computation_v1}
\title{Title}
\usage{
phenoscore_computation_v1(
  proteins_df,
  desired_phenotypes,
  pvalue_threshold = 0.05,
  path_length = 4,
  zscore_threshold = -1.96
)
}
\arguments{
\item{proteins_df}{dataframe containing up-reg or down-reg proteins}

\item{desired_phenotypes}{vector of interesting phenotypes}

\item{pvalue_threshold}{default 0.05}

\item{path_length}{default 4 (1 to 4 allowed)}

\item{zscore_threshold}{default -1.96}
}
\value{
a list of a plot and result table
}
\description{
Title
}
