% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore_network_preprocessing.R
\name{phenoscore_network_preprocessing}
\alias{phenoscore_network_preprocessing}
\title{Preprocess Network Data for Phenoscore Analysis}
\usage{
phenoscore_network_preprocessing(proteomics, phospho, local = FALSE)
}
\arguments{
\item{proteomics}{A dataframe containing processed proteomics data for SignalingProfiler.}

\item{phospho}{A dataframe containing processed phosphoproteomics data for SignalingProfiler.}

\item{local}{Logical. If `TRUE`, for *development purposes*, it uses a local path (`./inst/`) for the Python script;
otherwise, the function attempts to find the script in the package library path.}
}
\value{
A dataframe containing the cleaned and filtered SIGNOR interaction network.
}
\description{
This function processes proteomics and phosphoproteomics data to generate
a filtered interaction network from proteins to phenotypes based on expressed proteins.
It writes the input data to temporary files, runs a Python script
for processing, and returns the cleaned SIGNOR network.
}
\details{
The function writes the provided proteomics and phosphoproteomics data to TSV files,
executes a Python script for processing (available at `python/script.py`),
and retrieves the output. If the output file (`Global_result_final_table_minimized.txt`)
is successfully generated, it is read and returned
as a dataframe. If the process fails, an error is raised.
}
\examples{
data('prot_toy_df')
data('phospho_toy_df')

signor_network <- phenoscore_network_preprocessing(proteomics = prot_toy_df,
                                                   phospho = phospho_toy_df)
}
