% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phosphoscore_datasets}
\alias{phosphoscore_datasets}
\alias{good_phos_df_human_act}
\alias{good_phos_df_human_act_aapos}
\alias{good_phos_df_human_all}
\alias{good_phos_df_human_all_aapos}
\alias{good_phos_df_mouse_act}
\alias{good_phos_df_mouse_act_aapos}
\alias{good_phos_df_mouse_all}
\alias{good_phos_df_mouse_all_aapos}
\title{PhosphoScore - Regulatory Phosphosites Datasets}
\format{
A data frame with gene symbol, site ID, regulatory effect (activation or inhibition),
and optionally amino acid position.
Data frames with the following columns:
  \describe{
    \item{PHOSPHO_KEY_GN_SEQ}{Character, identifier of the phosphosite; `Gene-Sequence` or `Gene-Residue`}
    \item{UNIPROT}{Character, Target Protein Regulated by the TF}
    \item{ACTIVATION}{Character, Mode of Regulation (Activation/Inhibition)}
    \item{confidence}{Character, Confidence Score for Interaction (Present only in `tfea_db_*` datasets)}
  }

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6396 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6356 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9226 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9158 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2907 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2908 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5117 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5119 rows and 3 columns.
}
\source{
- <https://signor.uniroma2.it>
- <https://www.phosphosite.org>
}
\usage{
good_phos_df_human_act

good_phos_df_human_act_aapos

good_phos_df_human_all

good_phos_df_human_all_aapos

good_phos_df_mouse_act

good_phos_df_mouse_act_aapos

good_phos_df_mouse_all

good_phos_df_mouse_all_aapos
}
\description{
Curated regulatory phosphosites for human and mouse proteins from SIGNOR and PhosphoSitePlus.
Each dataset reports whether the phosphosite has an activating or inhibitory effect on the
activity or abundance of the substrate protein. Versions are provided with or without
amino acid position data.
}
\details{
These objects support the PhosphoScore pipeline for functional phosphoproteomics,
allowing integration of site-level regulatory knowledge. They differ by organism (human/mouse),
regulatory effect filtering (activity vs abundance and activity regulators), and presence of amino acid position.
}
\seealso{
\code{\link{get_phosphoscore_info}}
}
\keyword{datasets}
