% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkn_filtering_for_expression.R
\name{pkn_filtering_for_expression}
\alias{pkn_filtering_for_expression}
\title{Filters PKN for analytes present in experimental data}
\usage{
pkn_filtering_for_expression(
  PKN,
  transcriptomics,
  proteomics,
  phosphoproteomics
)
}
\arguments{
\item{PKN}{Igraph object containing the built-in Prior Knowledge Network}

\item{transcriptomics}{A tibble containing transcriptomics data}

\item{proteomics}{A tibble containing proteomics data}

\item{phosphoproteomics}{A tibble containing phosphoproteomics data}
}
\value{
igraph object containing only analytes present in experimental data
}
\description{
Filters PKN for analytes present in experimental data
}
