% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PKN_proteins}
\alias{PKN_proteins}
\alias{PKN_proteins_human}
\alias{PKN_proteins_human_atlas}
\alias{PKN_proteins_mouse}
\title{Molecular Entities in the Prior Knowledge Network (PKN)}
\format{
Each `PKN_proteins_*` object is a data frame with variable rows (depending on SignalingProfiler version) and 4 columns:
\describe{
  \item{ENTITY}{Character, Biological Entity Name}
  \item{ID}{Character, Biological Entity Identifier, from SIGNOR or UNIPROT}
  \item{TYPE}{Character, Biological Entity Molecular Type}
  \item{DATABASE}{Character, Database Source for Biological Entity Identifier}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9921 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8637 rows and 4 columns.
}
\source{
- <https://signor.uniroma2.it>
- <https://www.phosphosite.org>
- <https://pubmed.ncbi.nlm.nih.gov/36631611/> (for Ser/Thr kinase-substrate interactions)
- <https://pubmed.ncbi.nlm.nih.gov/38720073/> (for Tyr kinase-substrate interactions)
}
\usage{
PKN_proteins_human

PKN_proteins_human_atlas

PKN_proteins_mouse
}
\description{
Molecular entities connected in the Prior Knowledge Network (PKN) encompassing
all entities involved in *direct* and *indirect* interactions.
All `ENTITY` values are formatted consistently: uppercase for human,
and start-case for mouse protein names. Complexes and other `TYPE` are
always uppercase with special characters converted as underscores.
}
\details{
Available datasets:
- `PKN_proteins_human`: Molecular entities in the *human* Prior Knowledge Network.
- `PKN_proteins_human_atlas`: *Human* PKN entities, including kinase-substrate relationships.
- `PKN_proteins_mouse`: Molecular entities in the *mouse* Prior Knowledge Network.
}
\keyword{datasets}
