% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_paths_custom.R
\name{pkn_table_as_graph}
\alias{pkn_table_as_graph}
\title{PKN table as graph}
\usage{
pkn_table_as_graph(PKN_table)
}
\arguments{
\item{PKN_table}{table of PKN}
}
\value{
graph representing the PKN
}
\description{
PKN table as graph
}
