% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_PKN.R
\name{preprocess_PKN}
\alias{preprocess_PKN}
\title{preprocess_PKN}
\usage{
preprocess_PKN(omics_data, organism)
}
\arguments{
\item{omics_data}{list of all the omics data you have}

\item{organism}{string, 'mouse' or 'human'}
}
\value{
table of PKN with interactions involving expressed proteins in your system
}
\description{
preprocess_PKN
}
