% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prot_toy_df}
\alias{prot_toy_df}
\title{Proteomics toy dataset}
\format{
A data frame with one row per gene, and the following columns:
\describe{
  \item{gene_name}{Character. Gene symbol used as key in SignalingProfiler.}
  \item{UNIPROT}{Character. UniProt protein accession.}
  \item{difference}{Numeric. Fold-change in gene expression between the two conditions.}
  \item{logpval}{Numeric. -log10(p-value) of the statistical test assessing differential expression.}
  \item{significant}{Character. A '+' if the gene is significantly modulated, 'NA' otherwise.}
}
}
\usage{
prot_toy_df
}
\description{
A toy dataset representing proteomic differential expression results
from MCF7 cell lines upon metformin treatment (metformin vs control) (PMID: 27135362).
This dataset is structured to be compatible with SignalingProfiler.
}
\examples{
data('prot_toy_df')
}
\keyword{datasets}
