% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_information.R
\name{psp_phospho_parsing}
\alias{psp_phospho_parsing}
\title{psp_phospho_parsing}
\usage{
psp_phospho_parsing(organism, reg_site_path, only_activatory = TRUE)
}
\arguments{
\item{organism}{string, 'human' or 'mouse'}

\item{reg_site_path}{path to the PhosphoSitePlus 'Regulatory Sites' file}

\item{only_activatory}{Boolean, if TRUE, only interactions regulating protein activity are included}
}
\value{
table of regulatory phosphosites derived from PsP
}
\description{
psp_phospho_parsing
}
