% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_api.R
\name{query_uniprot_proteins}
\alias{query_uniprot_proteins}
\title{Retrieve UniProt Protein Information}
\usage{
query_uniprot_proteins(id_input, batch_size = 400)
}
\arguments{
\item{id_input}{Character vector. List of UniProt accession IDs.}

\item{batch_size}{Integer. Number of IDs per batch query to UniProt. Default: `400`.}
}
\value{
A `data.frame` containing UniProt metadata, including accession ID, protein name,
  Gene Name (Primary), organism, sequence length, and modification date.
}
\description{
Queries the UniProt REST API to retrieve protein metadata for a given set of UniProt IDs.
}
\examples{
result <- query_uniprot_proteins(id_input=c("P12345", "Q9Y6X1", "A0A0B4J2D5"))
print(result)

}
