% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore.R
\name{retrieve_functional_circuit}
\alias{retrieve_functional_circuit}
\title{retrieve_functional_circuit}
\usage{
retrieve_functional_circuit(SP_object, start_nodes, phenotype, k)
}
\arguments{
\item{SP_object}{sp_object list with network, nodes and edges}

\item{start_nodes}{character vector of nodes in the model}

\item{phenotype}{string reporting a phenotype}

\item{k}{integer, path length}
}
\value{
a subnetwork linking start nodes to phenotypes with a maximum of k length (shorter paths are included)
}
\description{
retrieve_functional_circuit
}
