% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{run_carnival_and_create_graph}
\alias{run_carnival_and_create_graph}
\title{Title}
\usage{
run_carnival_and_create_graph(
  source_df,
  target_df,
  naive_network,
  solver_path = NULL,
  solver = NULL,
  proteins_df,
  organism,
  prot_df,
  files = TRUE,
  path_sif = "./optimized_network.sif",
  path_rds = "./optimized_network.RDS"
)
}
\arguments{
\item{source_df}{df with source nodes discretized among 1 and -1}

\item{target_df}{df with target nodes in a continous range of activity}

\item{naive_network}{df with naive network interaction}

\item{solver_path}{path of the solver you use in CARNIVAL}

\item{solver}{name of solver}

\item{proteins_df}{all inferred proteins df}

\item{organism}{string, organism you are using}

\item{prot_df}{proteome df}

\item{files}{boolean value, if you want or not files}

\item{path_sif}{path of the sif output file of network}

\item{path_rds}{path of the rds output file of network}
}
\value{
list with igraph object, nodes df and edges df
}
\description{
Title
}
