% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{run_footprint_based_analysis}
\alias{run_footprint_based_analysis}
\title{run footprint based analysis}
\usage{
run_footprint_based_analysis(
  omic_data,
  analysis,
  organism,
  minsize,
  viper_significance,
  hypergeometric_correction
)
}
\arguments{
\item{omic_data}{phosphoproteomics or transcriptomics data}

\item{analysis}{tfea or ksea}

\item{organism}{mouse or human}

\item{minsize}{VIPER parameter for minsize of regulon}

\item{hypergeometric_correction}{if you want hypergeometric correction or not}
}
\value{
dataset of inferred proteins: tfs or kins and phos
}
\description{
run footprint based analysis
}
