% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{run_hypergeometric_test}
\alias{run_hypergeometric_test}
\title{Correct viper output with hypergeometric test}
\usage{
run_hypergeometric_test(omic_data, viper_output, analysis, organism)
}
\arguments{
\item{omic_data}{dataset containing experimental omic data}

\item{viper_output}{dataset containing viper enriched proteins}

\item{analysis}{type of analysis}

\item{organism}{organism}
}
\value{
table with weight for viper score correction
}
\description{
Correct viper output with hypergeometric test
}
