% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_databases.R
\name{signor_parsing}
\alias{signor_parsing}
\title{Parse SIGNOR Network Data}
\usage{
signor_parsing(
  organism,
  direct = FALSE,
  file_path = NULL,
  only_activatory = FALSE
)
}
\arguments{
\item{organism}{Character string, either `"human"` or `"mouse"`, specifying the organism.}

\item{direct}{Logical, whether to retain only direct interactions among biological entities; default `TRUE`.}

\item{file_path}{Character. Path to save the SIGNOR processed files. If `NULL`, results are not saved. Default: `NULL`.}

\item{only_activatory}{Logical. If `TRUE`, only interactions regulating protein activity are included. Default: `FALSE`.}
}
\value{
A list containing:
  - `proteins`: A data frame of SIGNOR entities (protein IDs and names).
  - `interactions`: A data frame of parsed SIGNOR interactions with regulatory effects.
}
\description{
Retrieves and processes interaction data from the SIGNOR database for the specified organism.
}
\details{
`only_activatory==TRUE`is used to create information for the PhosphoScore algorithm
}
\examples{
# Parse human SIGNOR interactions, including indirect interactions
signor_data <- signor_parsing("human")

# Parse mouse SIGNOR interactions, including only direct interactions
signor_data_direct <- signor_parsing("mouse", direct = TRUE)

# Save parsed results to a specific file path
signor_data_saved <- signor_parsing("human", file_path = "./", only_activatory = TRUE)
}
