% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_phenoscore_output}
\alias{toy_phenoscore_output}
\title{Toy Output of PhenoScore analysis (Step 3)}
\format{
A named list of 4 elements:
\describe{
  \item{barplot}{A \code{ggplot2} object representing metformin-induced phenotypes modulation as a barplot.}
  \item{table_regulators}{A data frame of 19 rows (10 phenotypes) with their predicted proteic activators and/or inhibitors in the network. Each phenotype can occur twice, one reporting activators and one regulators}
  \item{table_phenotypes}{A data frame of 10 rows (10 phenotypes) with their activity modulation `phenoscore` upon metformin treatment.}
  \item{sp_object_phenotypes}{A list reporting the metformin-induced signaling rewiring network (109 nodes and 309 edges), as *igraph_network*, *nodes_df*, and *edges_df*.}
}
}
\usage{
toy_phenoscore_output
}
\description{
A toy example of the result of phenotypes activity prediction with \code{SignalingProfiler}.
This object was obtained by applying `phenoscore_computation` function
on a multi-omic dataset derived from  MCF7 breast cancer cells treated with metformin.
The output includes graphical and tabular summaries
of predicted phenotype modulation based on signaling network analysis, and
(optionally) a network as *igraph object*.
}
\details{
This object illustrates the structure and content of the typical output generated by the \code{phenoscore_computation()} function in the SignalingProfiler workflow. It enables downstream visualization, interpretation, and pathway analysis of phenotype modulation in response to a perturbation—in this case, metformin treatment in MCF7 cells.
}
\examples{
data(toy_phenoscore_output)
names(toy_phenoscore_output)
toy_phenoscore_output$barplot
}
\seealso{
\code{\link{phenoscore_computation}}, \code{\link{toy_sp_output}}
}
\keyword{datasets}
