% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_prot_activity_df}
\alias{toy_prot_activity_df}
\title{Toy Output of Signaling Proteins Activity Inference (Step 1)}
\format{
A data frame of 92 proteins and 5 columns:
\describe{
  \item{`gene_name`}{The Gene Symbol of the inferred protein.}
  \item{`final_score`}{The combined activity score from both methods.}
  \item{`method`}{Indicates whether the score is derived from `"VIPER"`, `"PhosphoScore"`, or `"VIPER+PhosphoScore"`.}
  \item{`mf`}{Molecular function (`"tf"` for transcription factors, `"kin"` or `"phos"` for kinases and phosphatases).}
  \item{`UNIPROT`}{The UniProt ID of the inferred protein.}
}
}
\usage{
toy_prot_activity_df
}
\description{
A toy example of the result of protein activity prediction with \code{SignalingProfiler}.
This dataset contains 92 modulated proteins upon metformin treatment
inferred using `run_footprint_based_analysis` and `compute_phosphoscore`
from transcriptomics, proteomics, and phosphoproteomics datasets
of MCF7 cell lines upon metformin treatment (PMID:  27135362).
}
\examples{
data('toy_prot_activity_df')
head(toy_prot_activity_df)
}
\keyword{datasets}
