% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{weight_viper_score}
\alias{weight_viper_score}
\title{Title}
\usage{
weight_viper_score(ea_output)
}
\arguments{
\item{ea_output}{dataset representing ea output with viper and enrichment score}
}
\value{
dataset in input with weighted nes
}
\description{
Title
}
