% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{activity_from_proteomics}
\alias{activity_from_proteomics}
\title{Title}
\usage{
activity_from_proteomics(prot_df, organism)
}
\arguments{
\item{prot_df}{proteomic dataframe}

\item{organism}{string, 'human' or 'mouse'}
}
\description{
Title
}
