% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{add_output_carnival_nodes_attributes}
\alias{add_output_carnival_nodes_attributes}
\title{add_output_carnival_nodes_attributes}
\usage{
add_output_carnival_nodes_attributes(
  carnival_result,
  proteins_df,
  organism,
  with_atlas = TRUE,
  direct = FALSE
)
}
\arguments{
\item{carnival_result}{runCARNIVAL output}

\item{proteins_df}{inferred proteins df}

\item{organism}{string, 'mouse' or 'human'}

\item{with_atlas}{Boolean value, default TRUE, if FALSE exclude kinome atlas interactions}

\item{direct}{Boolean value, default FALSE, if TRUE uses direct interactions}
}
\value{
dataframe with nodes and attributes
}
\description{
add_output_carnival_nodes_attributes
}
