% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_carnival_attributes.R
\name{add_output_carnival_nodes_attributes}
\alias{add_output_carnival_nodes_attributes}
\title{Add Attributes to Nodes from CARNIVAL Optimization Output}
\usage{
add_output_carnival_nodes_attributes(
  carnival_result,
  proteins_df,
  organism,
  with_atlas = TRUE,
  direct = FALSE
)
}
\arguments{
\item{carnival_result}{A list returned from `run_carnival_and_create_graph()`,
\code{$nodesAttributes} and \code{$weightedSIF}.}

\item{proteins_df}{A data frame containing inferred protein information, including
columns such as \code{gene_name}, \code{UNIPROT}, \code{mf} (molecular function),
\code{final_score}, and \code{method}.}

\item{organism}{A character string, either \code{"mouse"} or \code{"human"},
indicating the species for which the prior knowledge network (PKN) should be used.}

\item{with_atlas}{Logical, default \code{TRUE}. If \code{FALSE}, kinome atlas
interactions are excluded.}

\item{direct}{Logical, default \code{FALSE}. If \code{TRUE}, uses only direct
interactions in the network.}
}
\value{
A \code{tibble} containing nodes and enriched attributes, including:
\describe{
  \item{gene_name}{Gene symbol of the node.}
  \item{carnival_activity}{Inferred activity score from CARNIVAL.}
  \item{UNIPROT}{UNIPROT identifier.}
  \item{mf}{Molecular function annotation.}
  \item{final_score}{Score derived from experimental data in Step 1.}
  \item{method}{Method used for protein inference.}
  \item{discordant}{Logical flag indicating inconsistency between
  inferred activity and final score.}
}
}
\description{
This function processes the nodes inferred by CARNIVAL and enriches them
with molecular attributes such as UNIPROT identifiers, molecular function scores,
and final activity scores. It also computes a `discordant` flag indicating
disagreement between inferred activity and experimental evidence.
}
\details{
Depending on the selected organism (human or mouse) and the source of prior knowledge
(e.g., with or without kinome atlas), it joins CARNIVAL output with relevant
protein databases and experimental data.
}
