% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{choose_database_for_building}
\alias{choose_database_for_building}
\title{choose_database_for_building}
\usage{
choose_database_for_building(organism, format)
}
\arguments{
\item{organism}{'human' or 'mouse''}

\item{format}{'igraph' or 'sif'}
}
\value{
the built in database you want
}
\description{
choose_database_for_building
}
