% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{choose_database_for_building}
\alias{choose_database_for_building}
\title{choose_database_for_building}
\usage{
choose_database_for_building(organism, format)
}
\arguments{
\item{organism}{'human' or 'mouse''}

\item{format}{'igraph', 'table'}
}
\value{
built-in database as igraph object or tibble
}
\description{
choose_database_for_building
}
