% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{choose_database_for_building}
\alias{choose_database_for_building}
\title{Get Prior Knowledge Network in multiple formats}
\usage{
choose_database_for_building(
  organism,
  with_atlas = FALSE,
  direct = FALSE,
  custom = FALSE,
  format
)
}
\arguments{
\item{organism}{Character string, either `"human"` or `"mouse"`, specifying the organism.}

\item{with_atlas}{Logical, whether to integrate inferred kinase-substrate
interactions from the Ser/Thr and Tyr Kinome Atlas (PMIDs: 36631611, 38720073); default: `FALSE`.}

\item{direct}{Logical, whether to retain only direct interactions among biological entities; default `TRUE`.}

\item{custom}{Logical, whether the user wants to provide a custom PKN.
If `TRUE`, the function will stop with an error message; default `FALSE`.}

\item{format}{Character string, format of the output; choose between `"igraph"` or `"table"`.}
}
\value{
The prior knowledge network in the specified format:
  - `"igraph"` → An `igraph` object.
  - `"table"` → A `data.frame`.
}
\description{
Returns the built-in Prior Knowledge Network (PKN) from SIGNOR and PhosphoSitePlus
in different formats, with optional kinase-substrate predicted interactions
from Kinome Atlases (PMIDs: 36631611, 38720073).
}
\details{
The function selects the appropriate dataset based on user-defined parameters.
If `custom = TRUE`, the function stops since custom PKNs must be handled separately.
For `"mouse"`, kinase-substrate interactions (`with_atlas = TRUE`) are not supported.
}
\examples{
# Retrieve human PKN in table format with direct interactions
choose_database_for_building(organism='human',with_atlas=FALSE,direct=TRUE,custom=FALSE,format='table')

# Retrieve mouse PKN as an igraph object (without kinase-substrate interactions)
PKN_mouse <- choose_database_for_building(organism = "mouse", format = "igraph")
}
