% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_PKN.R
\name{choose_PKN}
\alias{choose_PKN}
\title{Select a Prior Knowledge Network (PKN)}
\usage{
choose_PKN(
  organism,
  with_atlas = TRUE,
  direct = FALSE,
  custom = FALSE,
  custom_path = NULL
)
}
\arguments{
\item{organism}{Character. The organism for which to retrieve the PKN.
. Must be either `"human"` or `"mouse"`.}

\item{with_atlas}{Logical. If `TRUE`, integrates inferred kinase-substrate
interactions from the Ser/Thr and Tyr Kinome Atlas (PMIDs: 36631611, 38720073).
Applicable only to human PKNs. Default: `FALSE`.}

\item{direct}{Logical. If `TRUE`, retains only direct interactions among biological entities;
otherwise, includes indirect interactions. Default `TRUE`.}

\item{custom}{Logical. If `TRUE`, the user must provide a path to a custom
PKN file via the `custom_path` parameter. Default: `FALSE`.}

\item{custom_path}{Character. The file path of the custom PKN, expected
in tab-separated format. Must be provided if `custom = TRUE`. Default: `NULL`.}
}
\value{
A `data.frame` containing the selected Prior Knowledge Network (PKN).
  The columns may vary based on the chosen dataset but typically include
  source and target nodes, interaction types, and supporting evidence.
}
\description{
Retrieves a predefined or custom Prior Knowledge Network (PKN)
based on user-defined parameters. The PKN is a structured dataset
representing known interactions among signaling molecules.
}
\examples{
# Retrieve the human PKN without kinase-substrate interactions and with direct interactions
human_pkn <- choose_PKN(organism = "human", with_atlas = FALSE, direct = TRUE)

# Retrieve the mouse PKN with indirect interactions
mouse_pkn <- choose_PKN(organism = "mouse", direct = FALSE)

\dontrun{
# Load a custom PKN from a file
custom_pkn <- choose_PKN(organism = "human", custom = TRUE, custom_path = "your_pkn.tsv")
}
}
\seealso{
[PKN_human_dir], [PKN_mouse_dir]
}
