% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{combine_activityscore_proteoscore}
\alias{combine_activityscore_proteoscore}
\title{Title}
\usage{
combine_activityscore_proteoscore(activity_score, proteo_score)
}
\arguments{
\item{activity_score}{df with activity measures derived from phosphoproteomics}

\item{proteo_score}{df with activity measures derived from proteomics}
}
\value{
df with combined activity measures
}
\description{
Title
}
