% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{combine_footprint_and_phosphoscore}
\alias{combine_footprint_and_phosphoscore}
\title{Title}
\usage{
combine_footprint_and_phosphoscore(footprint_output, phosphoscore_df, analysis)
}
\arguments{
\item{footprint_output}{output of footprint based}

\item{phosphoscore_df}{output of phosphoscore based}

\item{analysis}{type of analysis}
}
\value{
dataset of inferred proteins with combined scores
}
\description{
Title
}
