% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore.R
\name{compute_phenoscore}
\alias{compute_phenoscore}
\title{compute_phenoscore}
\usage{
compute_phenoscore(sp_output, desired_phenotypes, compact = TRUE)
}
\arguments{
\item{sp_output}{Signaling Profiler output list}

\item{desired_phenotypes}{NULL if you want all phenotypes, vector of phenotypes name in uppercase}

\item{compact}{TRUE if you want a compact result, FALSE otherwise}
}
\value{
phenoSCORE table
}
\description{
compute_phenoscore
}
