% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_biological_ids.R
\name{convert_gene_name_in_uniprotid}
\alias{convert_gene_name_in_uniprotid}
\title{add uniprot id in dataset from gene name column}
\usage{
convert_gene_name_in_uniprotid(bio_dataset, organism)
}
\arguments{
\item{bio_dataset}{biological dataset without uniprot id}

\item{organism}{organism human or mouse}
}
\value{
biological dataset with uniprot id from SIGNOR added
}
\description{
add uniprot id in dataset from gene name column
}
