% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{convert_output_nodes_in_next_input}
\alias{convert_output_nodes_in_next_input}
\title{convert_output_nodes_in_next_input}
\usage{
convert_output_nodes_in_next_input(carnival_result)
}
\arguments{
\item{carnival_result}{}
}
\value{
formatted_nodes
}
\description{
convert_output_nodes_in_next_input
}
