% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_inferred_protein_PKN.R
\name{coverage_of_inferred_proteins_in_db}
\alias{coverage_of_inferred_proteins_in_db}
\title{Coverage of Inferred Proteins in Prior Knowledge Network}
\usage{
coverage_of_inferred_proteins_in_db(
  prediction_output,
  organism,
  with_atlas = TRUE,
  direct = FALSE,
  custom = FALSE,
  custom_path = FALSE,
  report = FALSE
)
}
\arguments{
\item{prediction_output}{`data.frame` containing inferred proteins with a column named `gene_name` and `mf` (molecular function).}

\item{organism}{Character. Either `'human'` or `'mouse'`.}

\item{with_atlas}{Logical. Default `TRUE`. If `TRUE`, considers PKN with Kinase-Substrate Atlas.}

\item{direct}{Logical. Default `FALSE`. If `TRUE`, considers PKN with only direct interactions.}

\item{custom}{Logical. Default `FALSE`. If `TRUE`, allows custom database selection.}

\item{custom_path}{Logical or character. Default `FALSE`. If a path is provided, it uses a custom PKN file.}

\item{report}{Logical. Default `FALSE`. If `TRUE`, writes results to a file (`report.txt`); otherwise, prints them.}
}
\value{
Prints coverage statistics or writes them to a file.
}
\description{
This function evaluates the presence of inferred proteins from omics data
in a selected Prior Knowledge Network (PKN) based on the chosen organism.
}
\examples{

inferred_proteins <- data.frame(
  gene_name = c("TP53", "EGFR", "MYC"),
  mf = c("tf", "kin", "rec")
)
# Control coverage in human PKN with direct interactions
coverage_of_inferred_proteins_in_db(inferred_proteins, organism = "human", with_atlas = F, direct = T)

# Control coverage in human PKN with indirect interactions
coverage_of_inferred_proteins_in_db(inferred_proteins, organism = "human", with_atlas = F, direct = F)

}
