% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{create_fasta}
\alias{create_fasta}
\title{Create fasta from phosphoproteomics files}
\usage{
create_fasta(phospho_df, path)
}
\arguments{
\item{phospho_df}{dataset of phosphoproteomics}

\item{path}{path of fasta file}
}
\value{
nothing
}
\description{
Create fasta from phosphoproteomics files
}
