% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_paths_custom.R
\name{create_graph_from_paths}
\alias{create_graph_from_paths}
\title{Convert Paths Data to an igraph Network}
\usage{
create_graph_from_paths(all_paths_df, PKN_table, connect_all = TRUE)
}
\arguments{
\item{all_paths_df}{Data frame, interactions in all shortest paths.}

\item{PKN_table}{Data frame, Prior Knowledge Network of causal (signed and oriented) interactions.}

\item{connect_all}{Logical, whether to connect intermediate nodes of shortest paths. Default: `TRUE`.}
}
\value{
An `igraph` object representing the network with all paths.
}
\description{
Converts a data frame of interactions (`all_paths_df`) into an `igraph` object,
optionally adding direct one-step connections between intermediate nodes.
}
