% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_paths_custom.R
\name{create_graph_from_paths}
\alias{create_graph_from_paths}
\title{create_graph_from_paths}
\usage{
create_graph_from_paths(all_paths_df, PKN_table)
}
\arguments{
\item{all_paths_df}{tibble with edges of shortest paths}

\item{PKN_table}{tibble of all causal interactions}
}
\value{
igraph object of naive network
}
\description{
create_graph_from_paths
}
