% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulons_creation.R
\name{create_ksea_regulons}
\alias{create_ksea_regulons}
\title{Generate Kinase-Substrate Enrichment Analysis (KSEA) Regulons}
\usage{
create_ksea_regulons(
  resources = c("SIGNOR", "PsP", "Omnipath"),
  organism,
  reg_site_path = NULL,
  kin_sub_path = NULL,
  omni_resources = c("PhosphoSite", "PhosphoSite_ProtMapper", "SIGNOR",
    "SIGNOR_ProtMapper", "Reactome_ProtMapper", "phosphoELM", "phosphoELM", "HPRD",
    "DEPOD"),
  local = FALSE
)
}
\arguments{
\item{resources}{A character vector specifying the regulatory databases to query.
Allowed values are `"SIGNOR"`, `"PsP"`, `"Omnipath"` and `"Atlas"` (to integrate Kinome-Atlas).
Default is `c("SIGNOR", "Omnipath")`.}

\item{organism}{Character string, either `"human"` or `"mouse"`, specifying the organism.}

\item{reg_site_path}{Character. Path to the PhosphoSitePlus "Regulatory_sites" file.}

\item{kin_sub_path}{Character. Path to the PhosphoSitePlus  "Kinases_Substrate_Dataset" file.}

\item{omni_resources}{A character vector specifying the OmniPath sources to use.
Default sources include `"PhosphoSite"`, `"SIGNOR"`, `"HPRD"`, `"DEPOD"`, and others.
For a complete list, refer to the OmniPath documentation.}

\item{local}{Logical. If `TRUE`, uses local package paths; otherwise, uses system-installed paths. Default: `FALSE`.}
}
\value{
A tibble containing the compiled kinase-substrate regulatory interactions with the following columns:
- `tf`: Kinase or phosphatase.
- `target`: Substrate, formatted as `Gene-Residue-Position` (e.g., `"AKT1-S-473"`).
- `mor`: Mode of regulation (`1` for phosphorylation, `-1` for dephosphorylation).
}
\description{
This function retrieves kinase-substrate interaction data from the SIGNOR and OmniPath databases,
creating a unified dataset of regulatory interactions for Kinase-Substrate Enrichment Analysis (KSEA).
}
\details{
- If `"SIGNOR"` is included in `resources`, the function retrieves kinase-substrate interactions from the SIGNOR database.
- If `"Omnipath"` is included, the function retrieves kinase-substrate interactions from the OmniPath database.
- The resulting table merges interactions from the selected resources into a single dataset.
}
\examples{
# Retrieve KSEA regulons from SIGNOR and OmniPath for human
ksea_regulons <- create_ksea_regulons(resources = c("SIGNOR", "Omnipath"), organism = "human")

# Retrieve KSEA regulons from OmniPath for mouse using specific resources
ksea_mouse <- create_ksea_regulons(resources = "Omnipath", organism = "mouse", omni_resources = c("SIGNOR", "DEPOD"))

}
