% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{create_matrix_from_VIPER_format}
\alias{create_matrix_from_VIPER_format}
\title{Create a matrix from VIPER format dataset}
\usage{
create_matrix_from_VIPER_format(viper_format)
}
\arguments{
\item{viper_format}{a dataset in VIPER format}
}
\value{
a matrix containing ID and t statistic
}
\description{
Create a matrix from VIPER format dataset
}
