% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_information.R
\name{create_tfea_regulons}
\alias{create_tfea_regulons}
\title{create_tfea_regulons}
\usage{
create_tfea_regulons(resources, organism)
}
\arguments{
\item{resources}{vector, allowed resources are SIGNOR, Dorothea or Collectri}

\item{organism}{string, 'human' and 'mouse'}
}
\value{
table of TF-target genes derived from user-defined resources
}
\description{
create_tfea_regulons
}
