% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{create_viper_format}
\alias{create_viper_format}
\title{Create VIPER format}
\usage{
create_viper_format(omic_data, analysis, significance)
}
\arguments{
\item{omic_data}{phosphoproteomics or transcriptomics dataset}

\item{analysis}{type of analysis ksea or tfea}

\item{significance}{boolean value (TRUE or FALSE) if you want all analytes
or just significant ones}
}
\value{
dataset in VIPER format
}
\description{
Create VIPER format
}
