% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{expand_and_map_edges}
\alias{expand_and_map_edges}
\title{expand_and_map_edges}
\usage{
expand_and_map_edges(
  optimized_graph_rds,
  optimized_graph_sif,
  organism,
  phospho_df,
  files,
  with_atlas = FALSE,
  path_sif,
  path_rds
)
}
\arguments{
\item{optimized_graph_rds}{igraph object output of the run_carnival_and_create_graph}

\item{optimized_graph_sif}{table output of the run_carnival_and_create_graph}

\item{organism}{#string, human or mouse}

\item{phospho_df}{tibble of phosphoproteomics data}

\item{files}{boolean value, TRUE if you want output files}

\item{with_atlas}{Boolean value, FALSE default, if TRUE uses integrated regulons}

\item{path_sif}{string of the path  of output sif file}

\item{path_rds}{string of the path of RDS file}
}
\value{
list with igraph object, optimized nodes df and optimized edges df
}
\description{
expand_and_map_edges
}
