% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{expand_and_map_edges}
\alias{expand_and_map_edges}
\title{expand_and_map_edges}
\usage{
expand_and_map_edges(
  optimized_object,
  organism,
  phospho_df,
  files,
  with_atlas = FALSE,
  direct = FALSE,
  path_sif,
  path_rds
)
}
\arguments{
\item{optimized_object}{SP object containing optimized graphs, nodes and edges of run_carnival_and_create_graph}

\item{organism}{#string, human or mouse}

\item{phospho_df}{tibble of phosphoproteomics data}

\item{files}{boolean value, TRUE if you want output files}

\item{with_atlas}{Boolean value, FALSE default, if TRUE uses integrated regulons}

\item{direct}{Boolean value, FALSE default, if TRUE uses indirect interactions for mapping}

\item{path_sif}{string of the path of output sif file (network)}

\item{path_rds}{string of the path of RDS file (SP object)}
}
\value{
list with igraph object, optimized nodes df and optimized edges df
}
\description{
expand_and_map_edges
}
