% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_carnival_nodes}
\alias{extract_carnival_nodes}
\title{Extract Nodes from an iGraph Object}
\usage{
extract_carnival_nodes(graph)
}
\arguments{
\item{graph}{An `igraph` object representing a network.}
}
\value{
A tibble with columns:
  \item{gene_name}{Node names in the network (renamed from `name`).}
  \item{UNIPROT}{UniProt accession IDs (if available in the graph).}
  \item{carnival_activity}{CARNIVAL activity scores (if available in the graph).}
}
\description{
This function extracts node (vertex) information from an `igraph` network
and formats it into a tibble with relevant attributes.
}
\examples{
library(igraph)
data('toy_opt_network')
extract_carnival_nodes(toy_opt_network$igraph_network)

}
