% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{filter_VIPER_output}
\alias{filter_VIPER_output}
\title{Filter VIPER output for molecular function}
\usage{
filter_VIPER_output(inferred_proteins_mf, analysis)
}
\arguments{
\item{inferred_proteins_mf}{dataset of inferred proteins with MF column}

\item{analysis}{string type of analysis}
}
\value{
dataset of inferred proteins with only the specific MF
}
\description{
Filter VIPER output for molecular function
}
