% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_paths_custom.R
\name{find_all_paths}
\alias{find_all_paths}
\title{find_all_paths}
\usage{
find_all_paths(v_start, v_end, PKN_table, max_length)
}
\arguments{
\item{v_start}{gene_name of starting node}

\item{v_end}{gene_name of target node}

\item{PKN_table}{tibble of all causal interactions}

\item{max_length}{max_length of the desired path, max = 4}
}
\value{
tibble of all shortest paths among start and end nodes
}
\description{
find_all_paths
}
