% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoscore.R
\name{generate_hybrid_db}
\alias{generate_hybrid_db}
\title{Generate a Hybrid Regulatory Phosphosite Database}
\usage{
generate_hybrid_db(mh_alignment, activatory)
}
\arguments{
\item{mh_alignment}{A dataframe of phosphosite alignments between mouse and human generated with [run_blast] function.}

\item{activatory}{Logical. If `TRUE`, considers only activatory sites.}
}
\value{
A dataframe containing hybrid regulatory phosphosites, mapped between species.
}
\description{
This function generates a hybrid regulatory phosphosite database
by aligning mouse phosphosites to human counterparts.
}
\examples{
# Example alignment dataset
alignment_data <- data.frame(
  q_ID = c("TP53-S15-mouse", "MYC-T22-mouse"),
  s_ID = c("TP53-S15-human", "MYC-T22-human"),
  UNIPROT = c("P04637", "P01106"),
  ACTIVATION = c(1, -1)
)

# Generate hybrid phosphosite database
hybrid_db <- generate_hybrid_db(alignment_data, activatory = TRUE)
head(hybrid_db)

}
\seealso{
run_blast()
}
