% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{generateTFList}
\alias{generateTFList}
\title{Generate Transcription Factor (TF) List}
\usage{
generateTFList(df = df, top = "all", access_idx = 1)
}
\arguments{
\item{df}{A dataframe of inferred proteins with their activity scores.}

\item{top}{An integer specifying the number of TFs to return (default: 50).
If `"all"`, returns all proteins in `df`.}

\item{access_idx}{An integer index referring to the column containing activity values.}
}
\value{
A list of ranked proteins with their scores.
}
\description{
Transform dataframe of proteins with their inferred activity in a list
compliant with R CARNIVAL analysis. Optionally, it extracts
the top-ranked proteins (TFs).
Function re-used from Saez-lab (https://saezlab.github.io/).
}
