% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{get_all_shortest_path_custom}
\alias{get_all_shortest_path_custom}
\title{get_all_shortest_path_custom}
\usage{
get_all_shortest_path_custom(start_nodes_gn, target_nodes_gn, db, path_length)
}
\arguments{
\item{start_nodes_gn}{vector of gene names of starting nodes}

\item{target_nodes_gn}{vector of gene names of target nodes}

\item{db}{igraph object representing chosen built in database}
}
\value{
a vector with edges ids along the shortest paths
}
\description{
get_all_shortest_path_custom
}
