% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_paths_custom.R
\name{get_all_shortest_path_custom}
\alias{get_all_shortest_path_custom}
\title{Find All Shortest Paths Between Sets of Nodes}
\usage{
get_all_shortest_path_custom(
  start_nodes_gn,
  target_nodes_gn,
  PKN_table,
  max_length
)
}
\arguments{
\item{start_nodes_gn}{Character vector, Gene Symbols of starting nodes.}

\item{target_nodes_gn}{Character vector, Gene Symbols of target nodes.}

\item{PKN_table}{Data frame, Prior Knowledge Network of causal (signed and oriented) interactions.}

\item{max_length}{Integer, maximum path length between start and end nodes (1 to 4).}
}
\value{
A data frame containing interactions in all shortest paths
  between the given sets of starting and target nodes.
}
\description{
Applies `find_all_paths` iteratively to find all paths between
multiple starting (`start_nodes_gn`) and target (`target_nodes_gn`) nodes.
}
\examples{
data('PKN_human_atlas_dir')
get_all_shortest_path_custom(
  start_nodes_gn = c("FLT3", "KRAS"),
  target_nodes_gn = c("STAT5A", "TFEB"),
  PKN_table = PKN_human_atlas_dir,
  max_length = 4)

}
