% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulons_creation.R
\name{get_omnipath}
\alias{get_omnipath}
\title{Retrieve Kinase-Substrate Interaction Data from OmniPath}
\usage{
get_omnipath(
  organism,
  resources = c("PhosphoSite", "PhosphoSite_ProtMapper", "SIGNOR", "SIGNOR_ProtMapper",
    "Reactome_ProtMapper", "phosphoELM", "HPRD", "DEPOD")
)
}
\arguments{
\item{organism}{Character string, either `"human"` or `"mouse"`, specifying the organism.}

\item{resources}{A character vector specifying the OmniPath resources to query.
Default sources include `"PhosphoSite"`, `"SIGNOR"`, `"Reactome_ProtMapper"`, `"HPRD"`, `"DEPOD"`, and others.
For a complete list, refer to the OmniPath documentation.}
}
\value{
A tibble containing kinase/phosphatase interactions with the following columns:
- `tf`: Kinase or phosphatase.
- `target`: Substrate, formatted as `Gene-Residue-Position` (e.g., `"MAPK1-T-185"`).
- `mor`: Mode of regulation (`1` for phosphorylation, `-1` for dephosphorylation).
}
\description{
This function queries the OmniPath database for kinase/phosphatase-substrate interactions,
retrieving phosphorylation and dephosphorylation sites for a specified organism.
}
\details{
- If no `resources` are specified, the function retrieves interactions from all available OmniPath resources.
- The function combines phosphorylation (`phosphorylation`) and dephosphorylation (`dephosphorylation`) interactions.
- The mode of regulation (`mor`) is assigned as `1` for phosphorylation and `-1` for dephosphorylation.
}
\examples{
# Retrieve OmniPath kinase-substrate interactions for human
omnipath_regulons <- get_omnipath(organism = "human")

# Retrieve interactions for mouse using specific resources
omnipath_mouse <- get_omnipath(organism = "mouse", resources = c("SIGNOR", "DEPOD"))
}
