% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulons_creation.R
\name{get_signor_regulons}
\alias{get_signor_regulons}
\title{Retrieve SIGNOR Regulons}
\usage{
get_signor_regulons(organism, analysis)
}
\arguments{
\item{organism}{Character string, either `"human"` or `"mouse"`, specifying the organism.}

\item{analysis}{A character string specifying the type of analysis. Can be `"tfea"` for transcriptional regulation or `"ksea"` for kinase-substrate interactions.}
}
\value{
A tibble containing the regulon data. The structure of the output depends on the analysis type:
- For `"tfea"`: Columns include `tf` (transcription factor), `target` (regulated gene), `mor` (mode of regulation), and `confidence` (set to `"A"`).
- For `"ksea"`: Columns include `tf` (kinase), `target` (substrate in `Gene-Residue-Position` format), and `mor` (mode of regulation).
}
\description{
This function extracts transcription factor regulons or kinase-substrate interactions from the SIGNOR database.
The function supports two types of analysis: Transcription Factor Enrichment Analysis (TFEA) and Kinase-Substrate Enrichment Analysis (KSEA).
}
\details{
- If `analysis = "tfea"`, the function retrieves transcriptional regulation interactions, returning a data frame of transcription factors and their target genes.
- If `analysis = "ksea"`, the function retrieves kinase-substrate phosphorylation interactions, returning a data frame of kinases and their phosphorylation targets.
}
\examples{
# Retrieve transcription factor regulons for human
tfea_regulons <- get_signor_regulons(organism = "human", analysis = "tfea")

# Retrieve kinase-substrate interactions for mouse
ksea_regulons <- get_signor_regulons(organism = "mouse", analysis = "ksea")

}
