% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{igraphToSif}
\alias{igraphToSif}
\title{igraphToSif}
\usage{
igraphToSif(inGraph, outfile = "output.sif", edgeLabel = "label")
}
\arguments{
\item{inGraph}{igraph object}

\item{outfile}{output file in sif format}

\item{edgeLabel}{label to use as edge}
}
\value{
igraph object in sif format
}
\description{
igraphToSif
}
