% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_databases.R
\name{integrate_atlas_in_pkn}
\alias{integrate_atlas_in_pkn}
\title{Integrate Ser/Thr Kinome Atlas and PhosphoSitePlus Data into PKN}
\usage{
integrate_atlas_in_pkn(
  reg_site_path,
  kin_sub_path,
  file_path = NULL,
  local = FALSE
)
}
\arguments{
\item{reg_site_path}{Character. Path to the PhosphoSitePlus "Regulatory_sites" file manually downloaded from it.}

\item{kin_sub_path}{Character. Path to the PhosphoSitePlus "Kinases_Substrate_Dataset" file manually downloaded from it.}

\item{file_path}{Character. Optional. Path to save the output file. If `NULL`, results are not saved. Default: `NULL`.}

\item{local}{Logical. If `TRUE`, uses local package paths; otherwise, uses system-installed paths. Default: `FALSE`.}
}
\value{
A `data.frame` containing integrated kinase-substrate interactions with
  regulatory effects mapped to SIGNOR conventions.
}
\description{
This function integrate kinase-substrate interactions from PhosphoSitePlus
and inferred in the Serine/Threonine and Tyrosine Kinome Atlases (PMIDs: 36631611, 38720073) into
a Prior Knowledge Network (PKN).
}
\details{
The function reads and processes:
- The **Kinome Atlas** dataset (filtered for 99% confidence that a kinase phosphorylate the substrate)
- The **PhosphoSitePlus Kinases_Substrates** file (manually downloaded)
- The **PhosphoSitePlus Regulatory_sites** file (manually downloaded)

The output includes interactions where phosphorylation sites have known functional effects
(e.g., activation or inhibition of substrates), mapped to the **SIGNOR** regulatory format.
}
\examples{
\dontrun{
# Example usage with local PhosphoSitePlus files
result <- integrate_atlas_in_pkn(
  reg_site_path = "./Regulatory_sites",
  kin_sub_path = "./Kinases_Substrate_Dataset",

# Save results to a specific file path
integrate_atlas_in_pkn(
  reg_site_path = "./Regulatory_sites",
  kin_sub_path = "./Kinases_Substrate_Dataset",
  file_path = "output/")
}
}
