% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_analysis.R
\name{keep_only_present_perturbation}
\alias{keep_only_present_perturbation}
\title{keep_only_present_perturbation}
\usage{
keep_only_present_perturbation(source_df, naive_network)
}
\arguments{
\item{source_df}{dataframe with receptors}

\item{naive_network}{dataframe with naive network interactions}
}
\value{
dataframe with receptors present in the naive network
}
\description{
keep_only_present_perturbation
}
