% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoscore.R
\name{map_experimental_on_regulatory_phosphosites}
\alias{map_experimental_on_regulatory_phosphosites}
\title{Map Experimental Phosphosites to Regulatory Phosphosites}
\usage{
map_experimental_on_regulatory_phosphosites(
  phosphoproteomic_data,
  organism,
  activatory,
  path_fasta,
  blastp_path = NULL,
  local = FALSE,
  custom = FALSE,
  custom_path = NULL
)
}
\arguments{
\item{phosphoproteomic_data}{A dataframe containing phosphoproteomics data.}

\item{organism}{Character string, either `"human"`, `"mouse"`, or `"hybrid"`, specifying the organism.
If  `"hybrid"` mouse regulatory information is remapped on human to enlarge coverage.}

\item{activatory}{Logical. If `TRUE`, considers only phosphosites regulating protein activity.
If `FALSE`, considers phosphosites regulating also protein abundance. Default: `TRUE`.}

\item{path_fasta}{A string specifying the path to the FASTA file for sequence mapping.}

\item{blastp_path}{Optional. A string specifying the Windows path of `blastp`.}

\item{local}{Logical. For **developmental purposes only**. Default: `FALSE`.}

\item{custom}{Logical. If `TRUE`, allows use of custom information on regulatory phosphosites.}

\item{custom_path}{A string specifying the path to a `.tsv` file containing custom  information on regulatory phosphosites.}
}
\value{
A list containing:
\describe{
  \item{`used_exp_data`}{A dataframe of experimental phosphosites that were successfully mapped to a regulatory phosphosite.}
  \item{`phosphoscore_df`}{A dataframe of inferred activity for each mapped phosphosite.}
}
}
\description{
Maps experimentally identified phosphosites onto
known regulatory phosphosites for activity inference.
}
\examples{
# Example phosphoproteomics dataset
phospho_data <- data.frame(
  gene_name = c("TP53", "MYC"),
  sequence_window = c("XXXXXXXXSXXXXXXX", "XXXXXXXTXXXXXXXX"),
  aminoacid = c("S", "T"),
  position = c(15, 22),
  significant = c("+", "+"),
  difference = c(1.2, -0.8)
)

# Map experimental phosphosites
mapped_results <- map_experimental_on_regulatory_phosphosites(phospho_data, "human", TRUE, "phospho.fasta")

}
