% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{map_experimental_on_regulatory_phosphosites}
\alias{map_experimental_on_regulatory_phosphosites}
\title{Title}
\usage{
map_experimental_on_regulatory_phosphosites(
  phosphoproteomic_data,
  organism,
  activatory,
  path_fasta,
  local = FALSE
)
}
\arguments{
\item{phosphoproteomic_data}{dataset of phosphoproteomics data}

\item{organism}{string, human mouse or hybrid}

\item{activatory}{boolean value, if relation activatory or all}

\item{local}{DEVELOPMENTAL PURPOSES TRUE OR FALSE}
}
\value{
phosphoscore_df representing experimentally quantified phosphosites
associated to their fold-change
}
\description{
Title
}
