% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_term_managment.R
\name{molecular_function_annotation}
\alias{molecular_function_annotation}
\title{Title}
\usage{
molecular_function_annotation(inferred_proteins_dataset, organism)
}
\arguments{
\item{inferred_proteins_dataset}{dataset of inferred proteins with UNIPROT}

\item{organism}{string, 'mouse', or 'human'}
}
\value{
dataset of inferred proteins with molecular function column
}
\description{
Title
}
