% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_information.R
\name{omnipath_parsing}
\alias{omnipath_parsing}
\title{omnipath_parsing}
\usage{
omnipath_parsing(resources, file_path = NULL)
}
\arguments{
\item{resources}{vector of resources (see OmniPath tutorial for the complete list)}

\item{file_path}{string, path for file saving}
}
\value{
table of interactions derived from OmniPath
}
\description{
omnipath_parsing
}
