% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{one_layer_naive_network}
\alias{one_layer_naive_network}
\title{one_layer_naive_network}
\usage{
one_layer_naive_network(
  starts_gn,
  targets_gn,
  PKN_table,
  max_length,
  rds_path = "one_layer_naive.RDS",
  sif_path = "one_layer_naive.sif"
)
}
\arguments{
\item{starts_gn}{gene names of starting nodes (e.g. receptors)}

\item{targets_gn}{gene names of target nodes (e.g. transcription factors)}

\item{PKN_table}{tibble of all causal interactions}

\item{max_length}{integer, 1 to 4 for max_length connecting start to end}

\item{rds_path}{path of network rds file}

\item{sif_path}{path of network sif file}
}
\value{
naive network
}
\description{
one_layer_naive_network
}
