% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore.R
\name{optimize_pheno_network}
\alias{optimize_pheno_network}
\title{optimize_pheno_network}
\usage{
optimize_pheno_network(
  sp_object,
  organism,
  phospho_df = NULL,
  carnival_options,
  files,
  direct = FALSE,
  with_atlas = FALSE,
  path_sif = "./pheno_opt_graph.sif",
  path_rds = "./pheno_opt_graph.rds"
)
}
\arguments{
\item{sp_object}{sp_object list with network, nodes and edges}

\item{organism}{string, human or mouse}

\item{phospho_df}{tibble of phosphoproteomics data}

\item{carnival_options}{list of options returned by default_CARNIVAL_options}

\item{files}{boolean value, TRUE if you want output files}

\item{direct}{Boolean value, default FALSE, if TRUE uses only direct interactions}

\item{with_atlas}{Boolean value, default TRUE, if FALSE excludes Kinome Altas derived regulons}

\item{path_sif}{path of the sif output file of network}

\item{path_rds}{path of the rds output file of network}
}
\value{
SP object list with igraph object, optimized nodes df and optimized edges df
}
\description{
optimize_pheno_network
}
