% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore.R
\name{pheno_to_start_circuit}
\alias{pheno_to_start_circuit}
\title{pheno_to_start_circuit}
\usage{
pheno_to_start_circuit(
  SP_object,
  start_nodes,
  phenotypes,
  k,
  start_to_top = FALSE
)
}
\arguments{
\item{SP_object}{sp_object list with network, nodes and edges}

\item{start_nodes}{character vector of nodes in the model}

\item{phenotypes}{vector of phenotypes}

\item{k}{integer, path length}

\item{start_to_top}{Boolean, to specify if you want to remove incoming edges on start nodes}
}
\value{
a subnetwork linking start nodes to phenotypes with a maximum of k length (shorter paths are included)
}
\description{
pheno_to_start_circuit
}
