% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore.R
\name{phenoscore_computation}
\alias{phenoscore_computation}
\title{phenoscore_computation}
\usage{
phenoscore_computation(
  proteins_df,
  desired_phenotypes = NULL,
  sp_graph,
  pheno_distances_table = NULL,
  path_length = 4,
  stat = "mean",
  zscore_threshold = -1.96,
  n_random = 1000,
  pvalue_threshold = 0.05,
  remove_cascade = TRUE,
  node_idx = FALSE,
  use_carnival_activity = FALSE
)
}
\arguments{
\item{proteins_df}{dataframe containing up-reg or down-reg proteins}

\item{desired_phenotypes}{vector of interesting phenotypes}

\item{sp_graph}{graph of signaling profiler output}

\item{path_length}{default 4 (1 to 4 allowed)}

\item{stat}{'mean' or 'median'}

\item{zscore_threshold}{default -1.96}

\item{n_random}{number of randomization, 1000}

\item{pvalue_threshold}{default 0.05}

\item{remove_cascade}{Boolean value, TRUE or FALSE}

\item{node_idx}{boolean}

\item{use_carnival_activity}{Boolean, TRUE or FALSE}

\item{phenoscore_distances_table}{df output of phenoscore network processing, in NULL take all}
}
\value{
a list of a plot and result table
}
\description{
phenoscore_computation
}
