% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscore.R
\name{phenoscore_network_preprocessing}
\alias{phenoscore_network_preprocessing}
\title{Title}
\usage{
phenoscore_network_preprocessing(proteomics, phospho)
}
\arguments{
\item{proteomics}{protemics dataframe processed for SP}

\item{phospho}{phosphoproteomics dataframe processed for SP}
}
\value{
SIGNOR cleaned according to expressed proteins
}
\description{
Title
}
