% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{phospho_score_hybrid_computation}
\alias{phospho_score_hybrid_computation}
\title{Title}
\usage{
phospho_score_hybrid_computation(
  phosphoproteomic_data,
  organism,
  activatory,
  blastp_path = NULL,
  path_fasta = "./phospho.fasta",
  local
)
}
\arguments{
\item{phosphoproteomic_data}{dataset containing experimental data}

\item{organism}{boolean value, if activatory or all interactions}

\item{blastp_path}{optional, path of Windows blastp exe file}

\item{path_fasta}{optional, path of phosphoproteomic fasta file}

\item{local}{}
}
\value{
list containing used experimental data and phosphoscore dataframe
}
\description{
Title
}
