% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoscore.R
\name{phosphoscore_computation}
\alias{phosphoscore_computation}
\title{Infer protein actiity from phosphosite regulatory power (PhophoScore)}
\usage{
phosphoscore_computation(
  phosphoproteomic_data,
  organism,
  activatory,
  path_fasta = "./phospho.fasta",
  blastp_path = NULL,
  local = FALSE,
  GO_annotation = FALSE,
  custom = FALSE,
  custom_path = NULL
)
}
\arguments{
\item{phosphoproteomic_data}{A dataframe containing phosphoproteomics measurements.}

\item{organism}{Character string, either `"human"`, `"mouse"`, or `"hybrid"`, specifying the organism.
If  `"hybrid"` mouse regulatory information is remapped on human to enlarge coverage.}

\item{activatory}{Logical. If `TRUE`, considers only phosphosites regulating protein activity.
If `FALSE`, considers phosphosites regulating also protein abundance. Default: `TRUE`.}

\item{path_fasta}{A string specifying the path to the FASTA file for `blastp` Necessary when `organism='hybrid'` Default: `'./phospho.fasta'`.}

\item{blastp_path}{Optional. A string specifying the Windows path of `blastp`.}

\item{local}{Logical. For **developmental purposes only**. Default: `FALSE`.}

\item{GO_annotation}{Logical indicating whether to perform GO molecular function annotation using [molecular_function_annotation].}

\item{custom}{Logical. If `TRUE`, allows use of custom information on regulatory phosphosites.}

\item{custom_path}{A string specifying the path to a `.tsv` file containing custom  information on regulatory phosphosites.}
}
\value{
A dataframe containing:
\describe{
  \item{`gene_name`}{The gene symbol.}
  \item{`phosphoscore`}{The inferred activity score.}
  \item{`n_sign_phos`}{Number of significant phosphosites used in scoring.}
  \item{`phos`}{Concatenated phosphosites contributing to the score.}
  \item{`act_rol`}{Activation or inhibition roles of each phosphosite.}
  \item{`phosphosite_fc`}{Fold-change values for the phosphosites.}
}
}
\description{
This function infers protein activity by combining phosphosites' significant
modulation in phosphoproteomics data with their regulatory role (*activation* or *inhibition*)
available in SIGNOR and PhosphoSitePlus databases. This activity measure
is called PhosphoScore.
}
\examples{
# Example phosphoproteomics dataset
phospho_data <- data.frame(
  gene_name = c("TP53", "MYC", "EGFR"),
  aminoacid = c("S", "T", "Y"),
  position = c(15, 22, 45),
  sequence_window = c("XXXXXXXXSXXXXXXX", "XXXXXXXTXXXXXXXX", "XXXXXXXYXXXXXXXX"),
  significant = c("+", "+", NA),
  difference = c(1.2, -0.8, 0.5)
)

# Compute PhosphoSCORE
phosphoscore_results <- phosphoscore_computation(phospho_data, "human", TRUE)
head(phosphoscore_results)

}
