% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{phosphoscore_computation}
\alias{phosphoscore_computation}
\title{Compute phosphoSCORE}
\usage{
phosphoscore_computation(
  phosphoproteomic_data,
  organism,
  activatory,
  path_fasta = "./phospho.fasta",
  blastp_path = NULL,
  local = FALSE,
  GO_annotation = FALSE
)
}
\arguments{
\item{phosphoproteomic_data}{dataset of phosphoproteomics measurments}

\item{organism}{string human or mouse}

\item{activatory}{boolean value}

\item{path_fasta}{optional}

\item{blastp_path}{optional, Windows path of blastp}

\item{local}{DEVELOPMENTAL PURPOSES}

\item{GO_annotation}{boolean value, TRUE perform GO molecular function annotaiton, FALSE default value}
}
\value{
phosphoscore dataset with gene_name, inferred activity and
used phosphosites from experimental data
}
\description{
Compute phosphoSCORE
}
