% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{phosphoscore_computation}
\alias{phosphoscore_computation}
\title{Compute phosphoSCORE}
\usage{
phosphoscore_computation(
  phosphoproteomic_data,
  organism,
  path_fasta = "./phospho.fasta"
)
}
\arguments{
\item{phosphoproteomic_data}{dataset of phosphoproteomics measurments}

\item{organism}{string human or mouse}
}
\value{
phosphoscore dataset with gene_name, inferred activity and
used phosphosites from experimental data
}
\description{
Compute phosphoSCORE
}
