% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{phosphoscore_computation_aapos}
\alias{phosphoscore_computation_aapos}
\title{phosphoscore_computation_aapos}
\usage{
phosphoscore_computation_aapos(
  phosphoproteomic_data,
  organism,
  activatory,
  GO_annotation = FALSE,
  custom = FALSE,
  custom_path = NULL
)
}
\arguments{
\item{phosphoproteomic_data}{table with SP formatted phosphoproteomic data}

\item{organism}{string, 'human' or 'mouse'}

\item{activatory}{Boolean, TRUE or FALSE}

\item{GO_annotation}{Boolean, TRUE or FALSE}
}
\value{
PhosphoScore table
}
\description{
phosphoscore_computation_aapos
}
