% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_PKN.R
\name{preprocess_PKN}
\alias{preprocess_PKN}
\title{preprocess_PKN}
\usage{
preprocess_PKN(omics_data, PKN_table)
}
\arguments{
\item{omics_data}{list of all the omics data you have}

\item{PKN_table}{table derived by choose_PKN}
}
\value{
table of PKN with interactions between proteins expressed in your system
}
\description{
preprocess_PKN
}
