% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_information.R
\name{psp_parsing}
\alias{psp_parsing}
\title{psp_parsing}
\usage{
psp_parsing(
  reg_site_path,
  kin_sub_path,
  organism,
  with_atlas = FALSE,
  file_path = NULL,
  local = FALSE
)
}
\arguments{
\item{reg_site_path}{path to the PhosphoSitePlus 'Regulatory Sites' file}

\item{kin_sub_path}{path to the PhosphoSitePlus 'Kinases-Substrates' file}

\item{organism}{string, 'human' or 'mouse'}

\item{with_atlas}{Boolean, if TRUE Ser/Thr Kinome Atlas interactions are included}

\item{file_path}{string, path for file saving}
}
\value{
table of interactions derived from PhosphoSitePlus
}
\description{
psp_parsing
}
