% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoscore.R
\name{run_blast}
\alias{run_blast}
\title{Perform BLAST Alignment on Phosphoproteomics Data}
\usage{
run_blast(
  path_experimental_fasta_file,
  all = FALSE,
  blastp_path = NULL,
  local = FALSE
)
}
\arguments{
\item{path_experimental_fasta_file}{A string specifying the path to the experimental FASTA file for `blastp`.}

\item{all}{Logical. If `TRUE`, returns all alignments; otherwise, only those with the same gene name.}

\item{blastp_path}{Optional. A string specifying the Windows path of `blastp`.}

\item{local}{Logical. For **developmental purposes only**. Default: `FALSE`.}
}
\value{
A list containing:
\describe{
  \item{`mapped`}{A dataframe of mouse phosphopeptides aligned to human phosphosites.}
  \item{`tocheck`}{A dataframe of ambiguous alignments (if `all = TRUE`).}
}
}
\description{
Runs BLAST to align experimental phosphoproteomics data against
a reference database of phosphopeptides.
In SignalingProfiler this function is used to map mouse phosphopeptides
on human phosphopeptides with annotated regulatory role
}
\examples{
\dontrun{
# Example: Running BLAST with an experimental FASTA file
blast_results <- run_blast("experiment.fasta", all = FALSE)
head(blast_results$mapped)
}

}
