% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{run_footprint_based_analysis}
\alias{run_footprint_based_analysis}
\title{run footprint based analysis}
\usage{
run_footprint_based_analysis(
  omic_data,
  analysis,
  organism,
  reg_minsize,
  exp_sign,
  integrated_regulons = FALSE,
  collectri = FALSE,
  hypergeom_corr,
  GO_annotation = FALSE
)
}
\arguments{
\item{omic_data}{dataset of experimental measured phosphosites or transcripts}

\item{analysis}{string, tfea or ksea}

\item{organism}{string, mouse or human}

\item{reg_minsize}{integer value, VIPER parameter for minsize of regulon}

\item{exp_sign}{boolean value, TRUE use only significant analytes,
FALSE: all measured analytes}

\item{integrated_regulons}{boolean value, default FALSE; if TRUE uses regulons derived from experimental data}

\item{hypergeom_corr}{boolean value, TRUE apply hypergeometric correction,
FALSE no correction}

\item{GO_annotation}{boolean value, TRUE perform GO molecular function annotaiton, FALSE default value}
}
\value{
dataset of inferred proteins: transcription factor (tfea)
or kinases and phosphatases (ksea)
}
\description{
run footprint based analysis
}
