% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{run_footprint_based_analysis}
\alias{run_footprint_based_analysis}
\title{run footprint based analysis}
\usage{
run_footprint_based_analysis(
  omic_data,
  analysis,
  organism,
  reg_minsize,
  exp_sign,
  integrated_regulons = FALSE,
  benchmark = NULL,
  collectri = FALSE,
  hypergeom_corr,
  correct_proteomics = FALSE,
  prot_df = NULL,
  GO_annotation = FALSE,
  custom = FALSE,
  custom_path = NULL
)
}
\arguments{
\item{omic_data}{dataset of experimental measured phosphosites or transcripts}

\item{analysis}{string, tfea or ksea}

\item{organism}{string, mouse or human}

\item{reg_minsize}{integer value, VIPER parameter for minsize of regulon}

\item{exp_sign}{boolean value, TRUE use only significant analytes,
FALSE: all measured analytes}

\item{integrated_regulons}{boolean value, default FALSE; if TRUE uses regulons derived from experimental data}

\item{benchmark}{string, default NULL, specify a string of a database for benchmarking}

\item{collectri}{boolean value, TRUE uses CollecTRI regulons}

\item{hypergeom_corr}{boolean value, TRUE apply hypergeometric correction,
FALSE no correction}

\item{correct_proteomics}{boolean value, TRUE correct VIPER output accordin to protein abundance}

\item{prot_df}{dataframe, containing proteomics data in sp format}

\item{GO_annotation}{boolean value, TRUE perform GO molecular function annotaiton, FALSE default value}
}
\value{
dataset of inferred proteins: transcription factor (tfea)
or kinases and phosphatases (ksea)
}
\description{
run footprint based analysis
}
