% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{run_hypergeometric_test}
\alias{run_hypergeometric_test}
\title{Correct viper output with hypergeometric test}
\usage{
run_hypergeometric_test(
  omic_data,
  viper_output,
  analysis,
  organism,
  integrated_regulons,
  collectri,
  custom = FALSE,
  custom_path = NULL
)
}
\arguments{
\item{omic_data}{dataset containing experimental omic data}

\item{viper_output}{dataset containing viper enriched proteins}

\item{analysis}{type of analysis}

\item{organism}{organism}

\item{integrated_regulons}{boolean,  if TRUE uses Kinome Atlas regulons}

\item{collectri}{boolean, if TRUE uses collectri regulons}

\item{custom}{Boolean, default FALSE, TRUE if you want to provide your regulons}

\item{custom_path}{string to the .tsv file of custom regulons}
}
\value{
table with weight for viper score correction
}
\description{
Correct viper output with hypergeometric test
}
