% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_activity_prediction.R
\name{run_viper}
\alias{run_viper}
\title{Run VIPER analysis}
\usage{
run_viper(
  viper_format,
  analysis,
  organism,
  reg_minsize,
  integrated_regulons = FALSE,
  collectri = FALSE,
  benchmark = NULL
)
}
\arguments{
\item{viper_format}{omic dataset in VIPER format}

\item{analysis}{string representing 'tfea' or 'ksea' analysis}

\item{organism}{string reporting the organism}

\item{reg_minsize}{viper function param: minimum regulon size to consider}

\item{integrated_regulons}{boolean value, default FALSE; if TRUE, uses regulons derived from experimental data}

\item{collectri}{boolean, if TRUE uses collectri regulons for tfea}

\item{benchmark}{string, default NULL, specifies a dataset used in tfea benchmarking}
}
\value{
a list containing significantly enriched proteins and
all inferred proteins
}
\description{
Run VIPER analysis
}
