% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_prediction.R
\name{run_viper}
\alias{run_viper}
\title{Run VIPER Analysis}
\usage{
run_viper(
  viper_format,
  analysis,
  organism,
  reg_minsize,
  integrated_regulons = FALSE,
  collectri = FALSE,
  custom = FALSE,
  custom_path = NULL
)
}
\arguments{
\item{viper_format}{A dataframe in VIPER-compatible format.}

\item{analysis}{String, specifying the type of analysis (`"tfea"` or `"ksea"`).}

\item{organism}{Character string, either `"human"` or `"mouse"`, specifying the organism.}

\item{reg_minsize}{Integer, minimum regulon size for VIPER.}

\item{integrated_regulons}{Logical, indicating whether to use regulons
derived from experimental data (Ser/Thr and Tyr Kinome Atlas (PMIDs: 36631611, 38720073)). Default: `FALSE`.}

\item{collectri}{Logical, indicating whether to use CollecTRI regulons
(only applicable to  `"tfea" `). Default: `FALSE`.}

\item{custom}{Logical indicating whether to use a custom regulon dataset. Default: `FALSE`.}

\item{custom_path}{String specifying the file path to a custom regulon dataset (if `custom = TRUE`). Default: `NULL`.}
}
\value{
A list containing:
  - `sign`: A dataframe of significantly enriched proteins.
  - `all`: A dataframe of all inferred proteins.
}
\description{
This function runs the VIPER analysis for transcription factor or kinase activity inference.
}
\examples{
# Run Transcription Factor Enrichment Analysis
viper_format <- create_viper_format(omic_data = tr_toy_df, analysis = 'tfea', significant = F)
results <- run_viper(viper_format = viper_format,
                     analysis = "tfea",
                     organism = "human",
                     reg_minsize = 10,
                     integrated_regulons = FALSE)
results$sign
results$all

}
