% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_information.R
\name{signor_parsing}
\alias{signor_parsing}
\title{signor_parsing}
\usage{
signor_parsing(
  organism,
  direct = FALSE,
  file_path = NULL,
  only_activatory = FALSE
)
}
\arguments{
\item{organism}{string, 'human' or 'mouse'}

\item{direct}{Boolean, if TRUE only direct interactions are included}

\item{file_path}{string, path for file saving}

\item{only_activatory}{Boolean, if TRUE, only interactions regulating protein activity are included}
}
\value{
list of SIGNOR entities and their interactions
}
\description{
signor_parsing
}
