% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{split_and_sort}
\alias{split_and_sort}
\title{Split and Sort Elements in a String}
\usage{
split_and_sort(x)
}
\arguments{
\item{x}{A character vector where each element consists of substrings
separated by underscores (`_`).}
}
\value{
A character vector where each element is sorted and unique,
  joined by underscores (`_`).
}
\description{
This function splits a string by underscores (`_`), sorts the elements
alphabetically, removes duplicates, and then concatenates them back
using an underscore separator.
}
\examples{
split_and_sort("B_A_C_A") # Returns "A_B_C"
split_and_sort(c("X_Y_Z", "C_A_B")) # Returns c("X_Y_Z", "A_B_C")
}
