% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{three_layer_naive_network}
\alias{three_layer_naive_network}
\title{three_layer_naive_network}
\usage{
three_layer_naive_network(
  receptors_gn,
  kinphos_gn,
  subs_gn,
  tfs_gn,
  db,
  rds_path = "three_layer_naive.RDS",
  sif_path = "three_layer_naive.sif"
)
}
\arguments{
\item{receptors_gn}{gene names of receptors}

\item{kinphos_gn}{gene names of kinases and phosphatases}

\item{subs_gn}{gene names of substrates}

\item{tfs_gn}{gene names of transcription factors}

\item{db}{built in causal relation database}
}
\value{
naive network
}
\description{
three_layer_naive_network
}
