% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_network_construction.R
\name{three_layer_naive_network}
\alias{three_layer_naive_network}
\title{three_layer_naive_network}
\usage{
three_layer_naive_network(
  starts_gn,
  intermediate1_gn,
  intermediate2_gn,
  targets_gn,
  PKN_table,
  max_length_1,
  max_length_2,
  max_length_3,
  both_intermediates = TRUE,
  keep_only_connected = FALSE,
  files = TRUE,
  rds_path = "three_layer_naive.RDS",
  sif_path = "three_layer_naive.sif",
  connect_all = FALSE
)
}
\arguments{
\item{starts_gn}{gene names of starting nodes}

\item{intermediate1_gn}{gene names of intermediates1 nodes}

\item{intermediate2_gn}{gene names of intermediates2 nodes}

\item{targets_gn}{gene names of targets nodes}

\item{PKN_table}{tibble of all causal interactions}

\item{max_length_1}{max_length of shortest path from start to intermediates1}

\item{max_length_2}{max_length of shortest path from intermediates1 to intermediates2}

\item{max_length_3}{max_length of shortest path from intermediates2 to targets}

\item{both_intermediates}{Boolean, TRUE if you want both intermediates as the starting point of the third layer,
FALSE if you want to use only intermediates2 (not suggested)}

\item{keep_only_connected}{Boolean, default FALSE, if TRUE keeps only intermediated connected}

\item{files}{Boolean, if true will save naive network files}

\item{rds_path}{path of network rds file}

\item{sif_path}{path of network sif file}

\item{connect_all}{if TRUE connect intermediate nodes}
}
\value{
naive network
}
\description{
three_layer_naive_network
}
