% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_prediction.R
\name{weight_viper_score}
\alias{weight_viper_score}
\title{Adjust VIPER Output Scores with Hypergeometric Test}
\usage{
weight_viper_score(ea_output)
}
\arguments{
\item{ea_output}{A dataframe containing VIPER output, including NES and enrichment scores.}
}
\value{
A dataframe with weighted NES scores for more refined protein activity inference.
}
\description{
This function adjusts the normalized enrichment scores (NES) from VIPER
using weights derived from a hypergeometric test.
}
\examples{

data('tr_toy_df')
viper_format <- create_viper_format(omic_data = tr_toy_df, analysis = 'tfea', significance = FALSE)
viper_output <- run_viper(viper_format = viper_format,
                     analysis = "tfea",
                     organism = "human",
                     reg_minsize = 10,
                     integrated_regulons = FALSE)
ea_output <- run_hypergeometric_test(tr_toy_df,
                                     viper_output$sign,
                                     "tfea",
                                     "human")
weighted_results <- weight_viper_score(ea_output=ea_output)

}
